Complexity measures API

Entropies.complexity — Function

complexity(c::ComplexityMeasure, x)

Estimate the complexity measure c for input data x, where c can be any of the following measures:

ReverseDispersion.

FuzzyEntropy.

Entropies.complexity_normalized — Function

complexity_normalized(c::ComplexityMeasure, x) → m ∈ [a, b]

Estimate the complexity measure c for input data x, normalized to the interval [a, b], where [a, b] depends on c.

Reverse dispersion entropy

Entropies.ReverseDispersion — Type

ReverseDispersion <: ComplexityMeasure
ReverseDispersion(; m = 2, τ = 1, check_unique = true,
    symbolization::SymbolizationScheme = GaussianSymbolization(c = 5)
)

Estimator for the reverse dispersion entropy complexity measure (Li et al., 2019)[Li2019].

Description

Li et al. (2021)[Li2019] defines the reverse dispersion entropy as

\[H_{rde} = \sum_{i = 1}^{c^m} \left(p_i - \dfrac{1}{{c^m}} \right)^2 = \left( \sum_{i=1}^{c^m} p_i^2 \right) - \dfrac{1}{c^{m}}\]

where the probabilities $p_i$ are obtained precisely as for the Dispersion probability estimator. Relative frequencies of dispersion patterns are computed using the given symbolization scheme , which defaults to symbolization using the normal cumulative distribution function (NCDF), as implemented by GaussianSymbolization, using embedding dimension m and embedding delay τ. Recommended parameter values[Li2018] are m ∈ [2, 3], τ = 1 for the embedding, and c ∈ [3, 4, …, 8] categories for the Gaussian mapping.

If normalizing, then the reverse dispersion entropy is normalized to [0, 1].

The minimum value of $H_{rde}$ is zero and occurs precisely when the dispersion pattern distribution is flat, which occurs when all $p_i$s are equal to $1/c^m$. Because $H_{rde} \geq 0$, $H_{rde}$ can therefore be said to be a measure of how far the dispersion pattern probability distribution is from white noise.

Data requirements

Like for Dispersion, the input must have more than one unique element for the default Gaussian mapping symbolization to be well-defined. Li et al. (2018) recommends that x has at least 1000 data points.

If check_unique == true (default), then it is checked that the input has more than one unique value. If check_unique == false and the input only has one unique element, then a InexactError is thrown when trying to compute probabilities.

Entropies.distance_to_whitenoise — Function

distance_to_whitenoise(p::Probabilities, estimator::ReverseDispersion; normalize = false)

Compute the distance of the probability distribution p from a uniform distribution, given the parameters of estimator (which must be known beforehand).

If normalize == true, then normalize the value to the interval [0, 1] by using the parameters of estimator.

Used to compute reverse dispersion entropy(ReverseDispersion; Li et al., 2019[Li2019]).

Fuzzy entropy

Entropies.FuzzyEntropy — Type

FuzzyEntropy([x]; r = 0.2std(x), kwargs...)

An estimator for the fuzzy entropy (Chen et al., 2007)[Chen2007] complexity/irregularity measure, used with complexity.

The keyword argument r is mandatory if an input timeseries x is not provided.

Keyword arguments

r::Real: The radius used when querying for nearest neighbors around points. Its value should be determined from the input data, for example as some proportion of the standard deviation of the data.

m::Int = 1: The embedding dimension.

n::Real = 2: The "fuzzy power". For d ∈ [-1, 1], the exponential function described below peaks at d == 0. For fixed r, larger n broadens the peak and smaller n narrows the peak.

τ::Int =: The embedding lag.

metric: The metric used to compute distances.

Description

Fuzzy entropy for an input timeseries x is computed as follows:

Construct two k-dimensional embeddings, one for k = m and one for k = m + 1, as follows:

\[{\bf x}_i^k = (x(i), x(i+τ), x(i+2τ), \ldots, x(i+(k-1)\tau)).\]

Zero-mean-normalize both embeddings.

Define the function

\[\phi(x, k, r, n) = \dfrac{1}{N - (m-1)\tau}\sum_{i = 1}^{N-(m-1)\tau} \dfrac{1}{N - (m-1)\tau - 1}\sum_{j = 1, j \neq i}^{N-(m-1)\tau} D_{ij}^m(n, r).\]

where $D_{ij}^m(n, r)$ is the "similarity degree" between vectors $\bf x_i ∈ \mathcal{R^k}$ and $\bf y_i ∈ \mathcal{R^k}$, given radius r, the "fuzzy power" n, and $D_{ij}^m$, which is the distance between $\bf x_i$ and $\bf y_i$ according to the given metric (Chen et al. (2007) uses the Chebyshev metric). The similarity degree is defined as

\[D_{ij}^m(n, r) = exp(-(d_{ij}^m)^n/r)\]

The fuzzy entropy is finally estimated as:

\[FuzzyEn(x, m, r, n) = \log(\phi(x, m, r, n)) - \log(\phi(x, m + 1, r, n))\]

Flexible embedding lag

In the original paper, they fix τ = 1. In our implementation, equations are modified to account for embeddings with τ != 1.

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.

Chen2007Chen, W., Wang, Z., Xie, H., & Yu, W. (2007). Characterization of surface EMG signal based on fuzzy entropy. IEEE Transactions on neural systems and rehabilitation engineering, 15(2), 266-272.