Histogram estimation

Entropies.non0hist — Method

non0hist(x::AbstractVector; normalize::Bool = true) → p::Vector{Float64}
non0hist(D::Dataset; normalize::Bool = true) → p::Vector{Float64}

Compute the unordered histogram of the values of x.

If normalize==true, then the histogram is sum-normalized. If normalize==false, then occurrence counts for the unique elements in x is returned. Assumes that x can be sorted.

Example

using Entropies
x = rand(1:10, 100000)
Entropies.non0hist(x) # sum-normalized
Entropies.non0hist(x, normalize = false) # histogram (counts)

using DelayEmbeddings, Entropies
D = Dataset(rand(1:3, 50000, 3))
Entropies.non0hist(D) # sum-normalized
Entropies.non0hist(D, normalize = false) # histogram (counts)