Multiscale

Multiscale API

The multiscale API is defined by the functions

multiscale

multiscale_normalized

downsample

which dispatch any of the MultiScaleAlgorithms listed below.

Missing docstring.

Missing docstring for MultiScaleAlgorithm. Check Documenter's build log for details.

ComplexityMeasures.Regular — Type

Regular <: MultiScaleAlgorithm
Regular(; f::Function = Statistics.mean)

The original multi-scale algorithm for multiscale entropy analysis (Costa et al., 2022)[Costa2002], which yields a single downsampled time series per scale s.

Description

Given a scalar-valued input time series x, the Regular multiscale algorithm downsamples and coarse-grains x by splitting it into non-overlapping windows of length s, and then constructing a new downsampled time series $D_t(s, f)$ by applying the function f to each of the resulting length-s windows.

The downsampled time series D_t(s) with t ∈ [1, 2, …, L], where L = floor(N / s), is given by:

\[\{ D_t(s, f) \}_{t = 1}^{L} = \left\{ f \left( \bf x_t \right) \right\}_{t = 1}^{L} = \left\{ {f\left( (x_i)_{i = (t - 1)s + 1}^{ts} \right)} \right\}_{t = 1}^{L}\]

where f is some summary statistic applied to the length-ts-((t - 1)s + 1) tuples xₖ. Different choices of f have yield different multiscale methods appearing in the literature. For example:

f == Statistics.mean yields the original first-moment multiscale sample entropy (Costa et al., 2002)[Costa2002].

f == Statistics.var yields the generalized multiscale sample entropy (Costa & Goldberger, 2015)[Costa2015], which uses the second-moment (variance) instead of the mean.

See also: Composite.

ComplexityMeasures.Composite — Type

Composite <: MultiScaleAlgorithm
Composite(; f::Function = Statistics.mean)

Composite multi-scale algorithm for multiscale entropy analysis (Wu et al., 2013)[Wu2013], used, with multiscale to compute, for example, composite multiscale entropy (CMSE).

Description

Given a scalar-valued input time series x, the composite multiscale algorithm, like Regular, downsamples and coarse-grains x by splitting it into non-overlapping windows of length s, and then constructing downsampled time series by applying the function f to each of the resulting length-s windows.

However, Wu et al. (2013)[Wu2013] realized that for each scale s, there are actually s different ways of selecting windows, depending on where indexing starts/ends. These s different downsampled time series D_t(s, f) at each scale s are constructed as follows:

\[\{ D_{k}(s) \} = \{ D_{t, k}(s) \}_{t = 1}^{L}, = \{ f \left( \bf x_{t, k} \right) \} = \left\{ {f\left( (x_i)_{i = (t - 1)s + k}^{ts + k - 1} \right)} \right\}_{t = 1}^{L},\]

where L = floor((N - s + 1) / s) and 1 ≤ k ≤ s, such that $D_{i, k}(s)$ is the i-th element of the k-th downsampled time series at scale s.

Finally, compute $\dfrac{1}{s} \sum_{k = 1}^s g(D_{k}(s))$, where g is some summary function, for example entropy or complexity.

Relation to Regular

The downsampled time series $D_{t, 1}(s)$ constructed using the composite multiscale method is equivalent to the downsampled time series $D_{t}(s)$ constructed using the Regular method, for which k == 1 is fixed, such that only a single time series is returned.

See also: Regular.

Missing docstring.

Missing docstring for downsample. Check Documenter's build log for details.

Missing docstring.

Missing docstring for multiscale. Check Documenter's build log for details.

Missing docstring.

Missing docstring for multiscale_normalized. Check Documenter's build log for details.

Available literature methods

A non-exhaustive list of literature methods, and the syntax to compute them, are listed below. Please open an issue or make a pull-request to ComplexityMeasures.jl if you find a literature method missing from this list, or if you publish a paper based on some new multiscale combination.

Method	Syntax	Reference

Refined composite multiscale dispersion entropy	multiscale(Composite(), Dispersion(), est, x, normalized = true)	Azami et al. (2017)[Azami2017]

Multiscale sample entropy (first moment)	multiscale(Regular(f = mean), SampleEntropy(), x)	Costa et al. (2002)[Costa2002]

Generalized multiscale sample entropy (second moment)	multiscale(Regular(f = std), SampleEntropy(), x)	Costa et al. (2015)[Costa2015]

Costa2002Costa, M., Goldberger, A. L., & Peng, C. K. (2002). Multiscale entropy analysis of complex physiologic time series. Physical review letters, 89(6), 068102.

Costa2015Costa, M. D., & Goldberger, A. L. (2015). Generalized multiscale entropy analysis: Application to quantifying the complex volatility of human heartbeat time series. EntropyDefinition, 17(3), 1197-1203.

Wu2013Wu, S. D., Wu, C. W., Lin, S. G., Wang, C. C., & Lee, K. Y. (2013). Time series analysis using composite multiscale entropy. Entropy, 15(3), 1069-1084.

Azami2017Azami, H., Rostaghi, M., Abásolo, D., & Escudero, J. (2017). Refined composite multiscale dispersion entropy and its application to biomedical signals. IEEE Transactions on Biomedical Engineering, 64(12), 2872-2879.

Costa2002Costa, M., Goldberger, A. L., & Peng, C. K. (2002). Multiscale entropy analysis of complex physiologic time series. Physical review letters, 89(6), 068102.

Costa2015Costa, M. D., & Goldberger, A. L. (2015). Generalized multiscale entropy analysis: Application to quantifying the complex volatility of human heartbeat time series. EntropyDefinition, 17(3), 1197-1203.