Utility methods

Fast histograms

Entropies.binhist — Function

binhist(x::AbstractDataset, ε::Real) → p, bins
binhist(x::AbstractDataset, ε::RectangularBinning) → p, bins

Hyper-optimized histogram calculation for x with rectangular binning ε. Returns the probabilities p of each bin of the histogram as well as the bins. Notice that bins are the starting corners of each bin. If ε isa Real, then the actual bin size is ε across each dimension. If ε isa RectangularBinning, then the bin size for each dimension will depend on the binning scheme.

See also: RectangularBinning.

Symbolization

Entropies.symbolize — Function

symbolize(x, scheme::SymbolizationScheme) → Vector{Int}
symbolize!(s, x, scheme::SymbolizationScheme) → Vector{Int}

Symbolize x using the provided symbolization scheme, optionally writing symbols into the pre-allocated symbol vector s. For usage examples, see individual symbolization scheme docstrings.

See also: OrdinalPattern, GaussianSymbolization.

Symbolization schemes

The following symbolization schemes are currently implemented.

Entropies.GaussianSymbolization — Type

GaussianSymbolization(; n_categories::Int = 5)

A symbolization scheme where the elements of x, which is the time series $x = \{x_j, j = 1, 2, \ldots, N \}$, is to a new time series $y = \{y_j, j = 1, 2, \ldots, N \}$ where $y_j \in (0, 1)$ and

\[y_j = \dfrac{1}{\sigma\sqrt{2\pi}} \int_{x = -\infty}^{x_j} \exp{\dfrac{-{(x-\mu)}^2}{2{\sigma}^2}} dx,\]

where $\mu$ and $\sigma$ are the mean and standard deviations of x.

Usage

symbolize(x::AbstractVector, s::GaussianSymbolization)

Map the elements of x to a symbol time series according to the Gaussian symbolization scheme s.

Examples

julia> x = [0.1, 0.4, 0.7, -2.1, 8.0, 0.9, -5.2];

julia> Entropies.symbolize(x, GaussianSymbolization(5))
7-element Vector{Int64}:
 3
 3
 3
 2
 5
 3
 1

See also: symbolize.

Entropies.OrdinalPattern — Type

OrdinalPattern(m = 3, τ = 1; lt = est.lt)

A symbolization scheme that converts the input time series to ordinal patterns, which are then encoded to integers using encode_motif.

Note

OrdinalPattern is intended for symbolizing time series. If providing a short vector, say x = [2, 5, 2, 1, 3, 4], then symbolize(x, OrdinalPattern(m = 2, τ = 1) will first embed x, then encode/symbolize each resulting state vector, not the original input. For symbolizing a single vector, use sortperm on it and use encode_motif on the resulting permutation indices.

Usage

symbolize(x, scheme::OrdinalPattern) → Vector{Int}
symbolize!(s, x, scheme::OrdinalPattern) → Vector{Int}

If applied to an m-dimensional Dataset x, then m and τ are ignored, and m-dimensional permutation patterns are obtained directly for each xᵢ ∈ x. Permutation patterns are then encoded as integers using encode_motif. Optionally, symbols can be written directly into a pre-allocated integer vector s, where length(s) == length(x) using symbolize!.

If applied to a univariate vector x, then x is first converted to a delay reconstruction using embedding dimension m and lag τ. Permutation patterns are then computed for each of the resulting m-dimensional xᵢ ∈ x, and each permutation is then encoded as an integer using encode_motif. If using the in-place variant with univariate input, s must obey length(s) == length(x)-(est.m-1)*est.τ.

Examples

using DelayEmbeddings, Entropies
D = Dataset([rand(7) for i = 1:1000])
s = symbolize(D, OrdinalPattern())

See also: symbolize.

Encoding ordinal patterns

Entropies.encode_motif — Function

encode_motif(x, m::Int = length(x)) → s::Int

Encode the length-m motif x (a vector of indices that would sort some vector v in ascending order) into its unique integer symbol $s \in \{1, 2, \ldots, m - 1 \}$, using Algorithm 1 in Berger et al. (2019)[Berger2019].

Example

v = rand(5)

# The indices that would sort `v` in ascending order. This is now a permutation
# of the index permutation (1, 2, ..., 5)
x = sortperm(v)

# Encode this permutation as an integer.
encode_motif(x)

Berger2019Berger, Sebastian, et al. "Teaching Ordinal Patterns to a Computer: Efficient Encoding Algorithms Based on the Lehmer Code." Entropy 21.10 (2019): 1023.