ComplexityMeasures.jl

ComplexityMeasures — Module

ComplexityMeasures.jl

A Julia package that provides:

a rigorous framework for extracting probabilities from data, based on the mathematical formulation of probability spaces

several (12+) outcome spaces, i.e., ways to discretize data into probabilities

several estimators for estimating probabilities given an outcome space, which correct theoretically known estimation biases

several definitions of information measures, such as various flavours of entropies (Shannon, Tsallis, Curado...), extropies, and probability-based complexity measures, that are used in the context of nonlinear dynamics, nonlinear timeseries analysis, and complex systems

several discrete and continuous (differential) estimators for entropies, which correct theoretically known estimation biases

estimators for other complexity measures that are not estimated based on probability functions

an extendable interface and well thought out API that makes it trivial to define new outcome spaces, or new estimators for probabilities, information measures, or complexity measures

ComplexityMeasures.jl can be used as a standalone package, or as part of other projects in the JuliaDynamics organization, such as DynamicalSystems.jl or CausalityTools.jl.

To install it, run import Pkg; Pkg.add("ComplexityMeasures").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Previously, this package was called Entropies.jl.

Content and terminology

Note

The documentation here was initially inspired by chapter 5 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

Before exploring the features of ComplexityMeasures.jl, it is useful to read through this terminology section. In the scientific literature, words like probabilities, entropy, and complexity measure are used (and abused) in multiple contexts, and are often used interchangeably to describe similar concepts. The API and documentation of ComplexityMeasures.jl aim to clarify the meaning and usage of these words, and to provide simple ways to obtain probabilities, information measures, or other complexity measures from input data.

For ComplexityMeasures.jl, we use the generic term "complexity measure" for any complexity measure that is not a direct functional of probabilities. "Information measure" is used exclusively about measures that are functionals of probability mass functions or probability density functions, for example entropies. We believe the general nonlinear dynamics community agrees with our take, as most papers that introduce different entropy flavors call them complexity measures. Example: "Permutation Entropy: A Natural Complexity Measure for Time Series" (Brandt and Pompe, 2002).

Probabilities and Outcome Spaces

Information measures, and some other complexity measures, are are typically computed based on probability distributions derived from input data. Probabilities can be either discrete (mass functions) or continuous (density functions). ComplexityMeasures.jl implements an interface for probabilities that exactly follows the mathematically rigorous formulation of probability spaces. In order to derive probabilities from data, an outcome space needs to be defined: a way to transform data into elements $\omega$ of an outcome space $\omega \in \Omega$, and assign probabilities to each outcome $p(\omega)$, such that $p(\Omega)=1$. $\omega$ are called outcomes or events.

If $\Omega$ is countable, this process is also called discretization of the input data. During discretization, each element of input data $\chi \in x$ is encoded (discretized) into an element of the outcome space $\omega \in \Omega$. Currently in ComplexityMeasures.jl, only countable $\Omega$ are implemented explicitly. Quantities that are estimated from probability density functions (i.e., uncountable $\Omega$) also exist. However, these are estimated by a one-step processes without the intermediate estimation of probabilities.

Once an outcome space (subtype of OutcomeSpace) is defined in ComplexityMeasures.jl, probabilities are estimated from input data using the probabilities function. The simplest way to do this is simply to assign a probability to each outcome as its relative count with respect to the total outcomes (i.e., normalizing the output of the counts function). This is also the default way. However, more advanced versions to estimate probabilities exist, that may account for known estimation biases. These probabilities estimators are subtypes of ProbabilitiesEstimator.

Information measures

Within ComplexityMeasures.jl, we have taken the pragmatic choice to label all measures that are explicit functionals of probability mass or probability density functions as information measures, even though they might not be labelled as information measures in the literature. This is encapsulated by the supertype InformationMeasure, whose subtypes are known (e.g., Shannon) and less-known (e.g., Curado) definitions of information measures, the most common of which are entropies.

But even "entropy" is an umbrella term that may mean several computationally, and sometimes even fundamentally, different quantities. In ComplexityMeasures.jl, we provide the generic function information that tries to both clarify disparate information-related measures, while unifying them under a common interface that highlights the modular nature of the term "information measure".

An information measure as defined by a subtype of InformationMeasure. However, estimating an information measure can be separated, on the highest level, into two main types:

Discrete information measures are functions of probability mass functions. Computing a discrete information measure boils down to two simple steps: defining an outcome space (using OutcomeSpace), then estimating a probability distribution (using a ProbabilitiesEstimator), then plugging the estimated probabilities into a discrete estimator of the information measure definition (DiscreteInfoEstimator).

Differential or continuous information measures are functions of probability density functions, which are integrals. Computing differential information measures therefore rely on estimating some density functional. For this task, we provide DifferentialInfoEstimators, which compute information measures via alternate means, without explicitly computing some probability distribution. For example, the Correa estimator computes the Shannon differential entropy using order statistics.

Crucially, many quantities in the nonlinear dynamics literature that are named as entropies, such as "permutation entropy" (entropy_permutation) and "wavelet entropy" (entropy_wavelet), are not really new entropies. They are the good old discrete Shannon entropy (Shannon), but calculated with new probabilities estimators. In turn, Shannon entropy is just one of many entropies and entropies are a subset of information measures. Nevertheless, we acknolwedge that names such as "permutation entropy" are commonplace, so in ComplexityMeasures.jl we provide convenience functions like entropy_permutation. However, we emphasize that these functions really aren't anything more than 2-lines-of-code wrappers that call information with the appropriate ProbabilitiesEstimator and InformationMeasure.

Other complexity measures

Other complexity measures which are not functionals of probability mass or density functions, yet still output some quantity related with complexity analysis, are called complexity measures within ComplexityMeasures.jl. They can be found in Complexity measures page. This includes measures like sample entropy and approximate entropy (which, even if named entropies, are not entropies in the formal mathematical sense).

A complexity measure is encapsulated by a ComplexityEstimator that can be given to the complexity function to obtain the corresponding numerical value. We stress that the separation between information measure and complexity measure is purely pragmatic, to establish a generic and extendable software interface within ComplexityMeasures.jl. One can always argue that "measure X should have been complexity and/or information" but in most cases this separation does not matter, it is only important that a classification choice is done and we stick with it.

Input data for ComplexityMeasures.jl

The input data type typically depend on the probability estimator chosen. In general though, the standard DynamicalSystems.jl approach is taken and as such we have three types of input data:

Timeseries, which are AbstractVector{<:Real}, used in e.g. with WaveletOverlap.

Multi-variate timeseries, or datasets, or state space sets, which are StateSpaceSets, used e.g. with NaiveKernel. The short syntax SSSet may be used instead of StateSpaceSet.

Spatial data, which are higher dimensional standard Arrays, used e.g. with SpatialSymbolicPermutation.

StateSpaceSets.StateSpaceSet — Type

StateSpaceSet{D, T} <: AbstractStateSpaceSet{D,T}

A dedicated interface for sets in a state space. It is an ordered container of equally-sized points of length D. Each point is represented by SVector{D, T}. The data are a standard Julia Vector{SVector}, and can be obtained with vec(ssset::StateSpaceSet). Typically the order of points in the set is the time direction, but it doesn't have to be.

When indexed with 1 index, StateSpaceSet is like a vector of points. When indexed with 2 indices it behaves like a matrix that has each of the columns be the timeseries of each of the variables. When iterated over, it iterates over its contained points. See description of indexing below for more.

StateSpaceSet also supports almost all sensible vector operations like append!, push!, hcat, eachrow, among others.

Description of indexing

In the following let i, j be integers, typeof(X) <: AbstractStateSpaceSet and v1, v2 be <: AbstractVector{Int} (v1, v2 could also be ranges, and for performance benefits make v2 an SVector{Int}).

X[i] == X[i, :] gives the ith point (returns an SVector)

X[v1] == X[v1, :], returns a StateSpaceSet with the points in those indices.

X[:, j] gives the jth variable timeseries (or collection), as Vector

X[v1, v2], X[:, v2] returns a StateSpaceSet with the appropriate entries (first indices being "time"/point index, while second being variables)

X[i, j] value of the jth variable, at the ith timepoint

Use Matrix(ssset) or StateSpaceSet(matrix) to convert. It is assumed that each column of the matrix is one variable. If you have various timeseries vectors x, y, z, ... pass them like StateSpaceSet(x, y, z, ...). You can use columns(dataset) to obtain the reverse, i.e. all columns of the dataset in a tuple.