Entropies.jl

Entropies — Module

Entropies.jl

A Julia package that provides estimators for probabilities, entropies, and other complexity emasures, in the context of nonlinear dynamics, nonlinear timeseries analysis, and complex systems. It can be used as a standalone package, or as part of several projects in the JuliaDynamics organization, such as DynamicalSystems.jl or CausalityTools.jl.

To install it, run import Pkg; Pkg.add("Entropies").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Info

You are reading the development version of the documentation of Entropies.jl, that will become version 2.0.

Terminology

Note

The documentation here follows (loosely) chapter 5 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

In the literature, the term "entropy" is used (and abused) in multiple contexts. The API and documentation of Entropies.jl aim to clarify some aspects of its usage, and to provide a simple way to obtain probabilities, entropies, or other complexity measures.

Probabilities

Entropies and other complexity measures are typically computed based on probability distributions. These can be obtained from input data in a plethora of different ways. The central API function that returns a probability distribution (or more precisely a probability mass function) is probabilities, which takes in a subtype of ProbabilitiesEstimator to specify how the probabilities are computed. All available estimators can be found in the estimators page.

Entropies

Entropy is an established concept in statistics, information theory, and nonlinear dynamics. However it is also an umbrella term that may mean several computationally, and sometimes even fundamentally, different quantities. In Entropies.jl, we provide the generic function entropy that tries to both clarify the disparate "entropy concepts", while unifying them under a common interface that highlights the modular nature of the word "entropy".

In the typical case, computing an entropy means computing a discrete entropy, which boils down to two simple steps: first estimating a probability distribution, and then applying one of the so-called "generalized entropy" definitions to the distributions.

A crucial thing to clarify is that in the nonlinear dynamics literature many quantities that are named as entropies (e.g., permutation entropy entropy_permutation, wavelet entropy entropy_wavelet, etc.), are not really new entropies. They are new probabilities estimators. They simply devise a new way to calculate probabilities from data, and then plug those probabilities into formal entropy formulas such as the Shannon entropy. The probabilities estimators are smartly created so that they elegantly highlight important aspects of the data relevant to complexity.

These names are commonplace, and so in Entropies.jl we provide convenience functions like entropy_wavelet. However, it should be noted that these functions really aren't anything more than 2-lines-of-code wrappers that call entropy with the appropriate ProbabilitiesEstimator.

What are genuinely different entropies are different definitions of entropy. And there are a lot of them! E.g., Renyi or Tsallis. These different definitions can be found in EntropyDefinition.

In addition to ProbabilitiesEstimators, we also provide DifferentialEntropyEstimators, which compute entropies via alternate means, without explicitly computing some probability distribution. For example, the Correa estimator computes Shannon differential entropy using order statistics. Differential entropies are functions of integrals, and usually rely on estimating some density functional.

Other complexity measures

Other complexity measures, which strictly speaking don't compute entropies, and may or may not explicitly compute probability distributions, are found in Complexity measures page. This includes measures like sample entropy and approximate entropy.

Input data for Entropies.jl

The input data type typically depend on the probability estimator chosen. In general though, the standard DynamicalSystems.jl approach is taken and as such we have three types of input data:

Timeseries, which are AbstractVector{<:Real}, used in e.g. with WaveletOverlap.

Multi-variate timeseries, or datasets, or state space sets, which are Dataset, used e.g. with NaiveKernel.

Spatial data, which are higher dimensional standard Arrays, used e.g. with SpatialSymbolicPermutation.

StateSpaceSets.Dataset — Type

Dataset{D, T} <: AbstractDataset{D,T}

A dedicated interface for datasets. It contains equally-sized datapoints of length D, represented by SVector{D, T}. These data are a standard Julia Vector{SVector}, and can be obtained with vec(dataset).

When indexed with 1 index, a dataset is like a vector of datapoints. When indexed with 2 indices it behaves like a matrix that has each of the columns be the timeseries of each of the variables.

Dataset also supports most sensible operations like append!, push!, hcat, eachrow, among others, and when iterated over, it iterates over its contained points.

Description of indexing

In the following let i, j be integers, typeof(data) <: AbstractDataset and v1, v2 be <: AbstractVector{Int} (v1, v2 could also be ranges, and for massive performance benefits make v2 an SVector{X, Int}).

data[i] == data[i, :] gives the ith datapoint (returns an SVector)

data[v1] == data[v1, :], returns a Dataset with the points in those indices.

data[:, j] gives the jth variable timeseries, as Vector

data[v1, v2], data[:, v2] returns a Dataset with the appropriate entries (first indices being "time"/point index, while second being variables)

data[i, j] value of the jth variable, at the ith timepoint

Use Matrix(dataset) or Dataset(matrix) to convert. It is assumed that each column of the matrix is one variable. If you have various timeseries vectors x, y, z, ... pass them like Dataset(x, y, z, ...). You can use columns(dataset) to obtain the reverse, i.e. all columns of the dataset in a tuple.