Entropies.jl

Entropies — Module

A Julia package that provides estimators for probabilities, entropies, and complexity measures for nonlinear dynamics, nonlinear timeseries analysis, and complex systems. It can be used as a standalone package, or as part of several projects in the JuliaDynamics organization, such as DynamicalSystems.jl or CausalityTools.jl.

To install it, run import Pkg; Pkg.add("Entropies").

API & terminology

Note

The documentation here follows (loosely) chapter 5 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

In the literature, the term "entropy" is used (and abused) in multiple contexts. The API and documentation of Entropies.jl aim to clarify some aspects of its usage, and to provide a simple way to obtain probabilities, entropies, or other complexity measures.

Probabilities

Entropies and other complexity measures are typically computed based on probability distributions. These are obtained from Input data for Entropies.jl in a plethora of different ways. The central API function that returns a probability distribution (in fact, just a vector of probabilities) is probabilities, which takes in a subtype of ProbabilitiesEstimator to specify how the probabilities are computed. All estimators available in Entropies.jl can be found in the estimators page.

Entropies

Entropy is an established concept in statistics, information theory, and nonlinear dynamics. However it is also an umbrella term that may mean several computationally different quantities. In Entropies.jl, we provide the generic function entropy that tries to both clarify the disparate "entropy concepts", while unifying them under a common interface that highlights the modular nature of the word "entropy".

Most of the time, computing an entropy boils down to two simple steps: first estimating a probability distribution, and then applying one of the so-called "generalized entropy" formulas to the distributions. Thus, any of the implemented probabilities estimators can be used to compute generalized entropies.

There aren't many entropies, really.

A crucial thing to clarify is that many quantities that are named as entropies (e.g., permutation entropy entropy_permutation, wavelet entropy entropy_wavelet, etc.), are not really new entropies. They are new probabilities estimators. They simply devise a new way to calculate probabilities from data, and then plug those probabilities into formal entropy formulas such as the Shannon entropy. The probabilities estimators are smartly created so that they elegantly highlight important aspects of the data relevant to complexity.

These names are common place, and so in Entropies.jl we provide convenience functions like entropy_wavelet. However, it should be noted that these functions really aren't anything more than 2-lines-of-code wrappers that call entropy with the appropriate ProbabilitiesEstimator.

There are only a few exceptions to this rule, which are quantities that are able to compute Shannon entropies via alternate means, without explicitly computing some probability distributions. These are IndirectEntropy instances, such as Kraskov.

Complexity measures

Other complexity measures, which strictly speaking don't compute entropies, and may or may not explicitly compute probability distributions, appear in the Complexity measures section.

Input data for Entropies.jl

The input data type typically depend on the probability estimator chosen. In general though, the standard DynamicalSystems.jl approach is taken and as such we have three types of input data:

Timeseries, which are AbstractVector{<:Real}, used in e.g. with WaveletOverlap.

Multi-dimensional timeseries, or datasets, or state space sets, which are Dataset, used e.g. with NaiveKernel.

Spatial data, which are higher dimensional standard Arrays, used e.g. with SpatialSymbolicPermutation.