Complexity measures API

Entropies.complexity — Function

complexity(c::ComplexityMeasure, x)

Estimate the complexity measure c for input data x, where c can be any of the following measures:

ReverseDispersion.

Entropies.complexity_normalized — Function

complexity_normalized(c::ComplexityMeasure, x) → m ∈ [a, b]

Estimate the complexity measure c for input data x, normalized to the interval [a, b], where [a, b] depends on c.

Reverse dispersion entropy

Entropies.ReverseDispersion — Type

ReverseDispersion <: ComplexityMeasure
ReverseDispersion(; m = 2, τ = 1, check_unique = true,
    symbolization::SymbolizationScheme = GaussianSymbolization(c = 5)
)

Estimator for the reverse dispersion entropy complexity measure (Li et al., 2019)[Li2019].

Description

Li et al. (2021)[Li2019] defines the reverse dispersion entropy as

\[H_{rde} = \sum_{i = 1}^{c^m} \left(p_i - \dfrac{1}{{c^m}} \right)^2 = \left( \sum_{i=1}^{c^m} p_i^2 \right) - \dfrac{1}{c^{m}}\]

where the probabilities $p_i$ are obtained precisely as for the Dispersion probability estimator. Relative frequencies of dispersion patterns are computed using the given symbolization scheme , which defaults to symbolization using the normal cumulative distribution function (NCDF), as implemented by GaussianSymbolization, using embedding dimension m and embedding delay τ. Recommended parameter values[Li2018] are m ∈ [2, 3], τ = 1 for the embedding, and c ∈ [3, 4, …, 8] categories for the Gaussian mapping.

If normalizing, then the reverse dispersion entropy is normalized to [0, 1].

The minimum value of $H_{rde}$ is zero and occurs precisely when the dispersion pattern distribution is flat, which occurs when all $p_i$s are equal to $1/c^m$. Because $H_{rde} \geq 0$, $H_{rde}$ can therefore be said to be a measure of how far the dispersion pattern probability distribution is from white noise.

Data requirements

Like for Dispersion, the input must have more than one unique element for the default Gaussian mapping symbolization to be well-defined. Li et al. (2018) recommends that x has at least 1000 data points.

If check_unique == true (default), then it is checked that the input has more than one unique value. If check_unique == false and the input only has one unique element, then a InexactError is thrown when trying to compute probabilities.

Entropies.distance_to_whitenoise — Function

distance_to_whitenoise(p::Probabilities, estimator::ReverseDispersion; normalize = false)

Compute the distance of the probability distribution p from a uniform distribution, given the parameters of estimator (which must be known beforehand).

If normalize == true, then normalize the value to the interval [0, 1] by using the parameters of estimator.

Used to compute reverse dispersion entropy(ReverseDispersion; Li et al., 2019[Li2019]).

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.