Complexity measures

Reverse dispersion entropy

Entropies.reverse_dispersion — Function

reverse_dispersion(x::AbstractVector{T}, est::Dispersion = Dispersion();
    normalize = true) where T <: Real

Compute the reverse dispersion entropy complexity measure (Li et al., 2019)[Li2019].

Description

Li et al. (2021)[Li2019] defines the reverse dispersion entropy as

\[H_{rde} = \sum_{i = 1}^{c^m} \left(p_i - \dfrac{1}{{c^m}} \right)^2 = \left( \sum_{i=1}^{c^m} p_i^2 \right) - \dfrac{1}{c^{m}}\]

where the probabilities $p_i$ are obtained precisely as for the Dispersion probability estimator. Relative frequencies of dispersion patterns are computed using the given symbolization scheme , which defaults to symbolization using the normal cumulative distribution function (NCDF), as implemented by GaussianSymbolization, using embedding dimension m and embedding delay τ. Recommended parameter values[Li2018] are m ∈ [2, 3], τ = 1 for the embedding, and c ∈ [3, 4, …, 8] categories for the Gaussian mapping.

If normalize == true, then the reverse dispersion entropy is normalized to [0, 1].

The minimum value of $H_{rde}$ is zero and occurs precisely when the dispersion pattern distribution is flat, which occurs when all $p_i$s are equal to $1/c^m$. Because $H_{rde} \geq 0$, $H_{rde}$ can therefore be said to be a measure of how far the dispersion pattern probability distribution is from white noise.

Entropies.distance_to_whitenoise — Function

distance_to_whitenoise(p::Probabilities, estimator::Dispersion; normalize = false)

Compute the distance of the probability distribution p from a uniform distribution, given the parameters of estimator (which must be known beforehand).

If normalize == true, then normalize the value to the interval [0, 1] by using the parameters of estimator.

Used to compute reverse dispersion entropy(reverse_dispersion; Li et al., 2019[Li2019]).

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.

Li2019Li, Y., Gao, X., & Wang, L. (2019). Reverse dispersion entropy: a new complexity measure for sensor signal. Sensors, 19(23), 5203.