Entropies

Rényi (generalized) entropy

Entropies.entropy_renyi — Function

entropy_renyi(p::Probabilities; q = 1.0, base = MathConstants.e)

Compute the Rényi[Rényi1960] generalized order-q entropy of some probabilities (typically returned by the probabilities function).

entropy_renyi(x::Array_or_Dataset, est; q = 1.0, base)

A convenience syntax, which calls first probabilities(x, est) and then calculates the entropy of the result (and thus est can be anything the probabilities function accepts).

Description

Let $p$ be an array of probabilities (summing to 1). Then the Rényi generalized entropy is

\[H_q(p) = \frac{1}{1-q} \log \left(\sum_i p[i]^q\right)\]

and generalizes other known entropies, like e.g. the information entropy ($q = 1$, see [Shannon1948]), the maximum entropy ($q=0$, also known as Hartley entropy), or the correlation entropy ($q = 2$, also known as collision entropy).

See also: maxentropy_renyi.

Tsallis (generalized) entropy

Entropies.entropy_tsallis — Function

entropy_tsallis(p::Probabilities; k = 1, q = 0, base = MathConstants.e)

Compute the Tsallis entropy of p (Tsallis, 1998)[Tsallis1988].

base only applies in the limiting case q == 1, in which the Tsallis entropy reduces to Shannon entropy.

entropy_tsallis(x::Array_or_Dataset, est; k = 1, q = 0, base = MathConstants.e)

A convenience syntax, which calls first probabilities(x, est) and then calculates the Tsallis entropy of the result (and thus est can be anything the probabilities function accepts).

Description

The Tsallis entropy is a generalization of the Boltzmann-Gibbs entropy, with k standing for the Boltzmann constant. It is defined as

\[S_q(p) = \frac{k}{q - 1}\left(1 - \sum_{i} p[i]^q\right)\]

Shannon entropy (convenience)

Entropies.entropy_shannon — Function

entropy_shannon(args...; base = MathConstants.e)

Equivalent to entropy_renyi(args...; base = base, q = 1) and provided solely for convenience. Computes the Shannon entropy, given by

\[H(p) = - \sum_i p[i] \log(p[i])\]

See also: maxentropy_shannon.

Normalization

The generic entropy_normalized normalizes any entropy value to the entropy of a uniform distribution. We also provide maximum entropy functions that are useful for manual normalization.

Entropies.entropy_normalized — Function

entropy_normalized(f::Function, x, est::ProbabilitiesEstimator, args...; kwargs...)

Convenience syntax for normalizing to the entropy of uniform probability distribution. First estimates probabilities as p::Probabilities = f(x, est, args...; kwargs...), then calls entropy_normalized(f, p, args...; kwargs...).

Normalization is only defined for estimators for which alphabet_length is defined, meaning that the total number of states or symbols is known beforehand.

entropy_normalized(f::Function, p::Probabilities, est::ProbabilitiesEstimator, args...;
    kwargs...)

Normalize the entropy, as returned by the entropy function f called with the given arguments (i.e. f(p, args...; kwargs...)), to the entropy of a uniform distribution, inferring alphabet_length from est.

entropy_normalized(f::Function, p::Probabilities, args...; kwargs...)

The same as above, but infers alphabet length from counting how many elements are in p (zero probabilities are counted).

Examples

Computing normalized entropy from scratch:

x = rand(100)
entropy_normalized(entropy_renyi, x, Dispersion())

Computing normalized entropy from pre-computed probabilities with known parameters:

x = rand(100)
est = Dispersion(m = 3, symbolization = GaussianSymbolization(c = 4))
p = probabilities(x, est)
entropy_normalized(entropy_renyi, p, est)

Computing normalized entropy, assumming there are N = 10 total states:

N = 10
p = Probabilities(rand(10))
entropy_normalized(entropy_renyi, p, est)

Normalized output range

For Rényi entropy (e.g. Kumar et al., 1986), and for Tsallis entropy (Tsallis, 1998), normalizing to the uniform distribution ensures that the entropy lies in the interval [0, 1]. For other entropies and parameter choices, the resulting entropy is not guaranteed to lie in [0, 1]. It is up to the user to decide whether normalizing to a uniform distribution makes sense for their use case.

Indirect entropies

Here we list functions which compute Shannon entropies via alternate means, without explicitly computing some probability distributions and then using the Shannon formulat.

Nearest neighbors entropy

Entropies.entropy_kraskov — Function

entropy_kraskov(x::AbstractDataset{D, T}; k::Int = 1, w::Int = 0,
    base::Real = MathConstants.e) where {D, T}

Estimate Shannon entropy to the given base using k-th nearest neighbor searches (Kraskov, 2004)[Kraskov2004].

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

See also: entropy_kozachenkoleonenko.

Entropies.entropy_kozachenkoleonenko — Function

entropy_kozachenkoleonenko(x::AbstractDataset{D, T}; k::Int = 1, w::Int = 0,
    base::Real = MathConstants.e) where {D, T}

Estimate Shannon entropy to the given base using k-th nearest neighbor searches, using the method from Kozachenko & Leonenko (1987)[KozachenkoLeonenko1987], as described in Charzyńska and Gambin (2016)[Charzyńska2016].

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

See also: entropy_kraskov.

Convenience functions

In this subsection we expand documentation strings of "entropy names" that are used commonly in the literature, such as "permutation entropy". As we made clear in API & terminology, these are just the existing Shannon/Rényi/Tsallis entropy with a particularly chosen probability estimator. We have only defined convenience functions for the most used names, and arbitrary more specialized convenience functions can be easily defined in a couple lines of code.

Entropies.entropy_permutation — Function

entropy_permutation(x; τ = 1, m = 3, base = MathConstants.e)

Compute the permutation entropy of order m with delay/lag τ. This function is just a convenience call to:

est = SymbolicPermutation(; m, τ)
entropy_shannon(x, est; base)

See SymbolicPermutation for more info. Similarly, one can use SymbolicWeightedPermutation or SymbolicAmplitudeAwarePermutation for the weighted/amplitude-aware versions.

Entropies.entropy_spatial_permutation — Function

entropy_spatial_permutation(x, stencil, periodic = true; kwargs...)

Compute the spatial permutation entropy of x given the stencil. Here x must be a matrix or higher dimensional Array containing spatial data. This function is just a convenience call to:

est = SpatialSymbolicPermutation(stencil, x, periodic)
entropy_shannon(x, est; kwargs...)

See SpatialSymbolicPermutation for more info, or how to encode stencils.

Entropies.entropy_wavelet — Function

entropy_wavelet(x; wavelet = Wavelets.WT.Daubechies{12}(), base = MathConstants.e)

Compute the wavelet entropy. This function is just a convenience call to:

est = WaveletOverlap(wavelet)
entropy_shannon(x, est; base)

See WaveletOverlap for more info.

Entropies.entropy_dispersion — Function

entropy_dispersion(x; m = 2, τ = 1, s = GaussianSymbolization(3),
    base = MathConstants.e)

Compute the dispersion entropy. This function is just a convenience call to:

est = Dispersion(m = m, τ = τ, s = s)
entropy_shannon(x, est; base)

See Dispersion for more info.

Rényi1960A. Rényi, Proceedings of the fourth Berkeley Symposium on Mathematics, Statistics and Probability, pp 547 (1960)

Kumar1986Kumar, U., Kumar, V., & Kapur, J. N. (1986). Normalized measures of entropy. International Journal Of General System, 12(1), 55-69.

Shannon1948C. E. Shannon, Bell Systems Technical Journal 27, pp 379 (1948)

Tsallis1988Tsallis, C. (1988). Possible generalization of Boltzmann-Gibbs statistics. Journal of statistical physics, 52(1), 479-487.

Kumar1986Kumar, U., Kumar, V., & Kapur, J. N. (1986). Normalized measures of entropy. International Journal Of General System, 12(1), 55-69.

Tsallis1998Tsallis, C. (1988). Possible generalization of Boltzmann-Gibbs statistics. Journal of statistical physics, 52(1), 479-487.

Kraskov2004Kraskov, A., Stögbauer, H., & Grassberger, P. (2004). Estimating mutual information. Physical review E, 69(6), 066138.

Charzyńska2016Charzyńska, A., & Gambin, A. (2016). Improvement of the k-NN entropy estimator with applications in systems biology. Entropy, 18(1), 13.

KozachenkoLeonenko1987Kozachenko, L. F., & Leonenko, N. N. (1987). Sample estimate of the entropy of a random vector. Problemy Peredachi Informatsii, 23(2), 9-16.