Entropies

Entropies.entropy — Function

entropy([e::Entropy,] x, est::ProbabilitiesEstimator) → h::Real
entropy([e::Entropy,] probs::Probabilities) → h::Real

Compute a (generalized) entropy h from x according to the specified entropy type e and the given probability estimator est. Alternatively compute the entropy directly from the existing probabilities probs. In fact, the first method is a 2-lines-of-code wrapper that calls probabilities and gives the result to the second method.

x is typically an Array or a Dataset, see Input data for Entropies.jl.

The entropy types that support this interface are "direct" entropies. They always yield an entropy value given a probability distribution. Such entropies are theoretically well-founded and are typically called "generalized entropies". Currently implemented types are:

Renyi.

Tsallis.

Shannon, which is a subcase of the above two in the limit q → 1.

Curado.

StretchedExponential.

The entropy (first argument) is optional: if not given, Shannon() is used instead.

These entropies also have a well defined maximum value for a given probability estimator. To obtain this value one only needs to call the entropy_maximum function with the chosen entropy type and probability estimator. Or, one can use entropy_normalized to obtain the normalized form of the entropy (divided by the maximum).

Examples

x = [rand(Bool) for _ in 1:10000] # coin toss
ps = probabilities(x) # gives about [0.5, 0.5] by definition
h = entropy(ps) # gives 1, about 1 bit by definition
h = entropy(Shannon(), ps) # syntactically equivalent to above
h = entropy(Shannon(), x, CountOccurrences()) # syntactically equivalent to above
h = entropy(x, SymbolicPermutation(;m=3)) # gives about 2, again by definition
h = entropy(Renyi(2.0), ps) # also gives 1, order `q` doesn't matter for coin toss

entropy(e::IndirectEntropy, x) → h::Real

Compute the entropy of x, here named h, according to the specified indirect entropy estimator e.

In contrast to the "typical" way one obtains entropies in the above methods, indirect entropy estimators compute Shannon entropies via alternate means, without explicitly computing probability distributions. The available indirect entropies are:

Kraskov.

KozachenkoLeonenko.

Vasicek.

Entropies.entropy! — Function

entropy!(s, [e::Entropy,] x, est::ProbabilitiesEstimator)

Similar to probabilities!, this is an in-place version of entropy that allows pre-allocation of temporarily used containers.

The entropy (second argument) is optional: if not given, Shannon() is used instead.

Only works for certain estimators. See for example SymbolicPermutation.

Rényi (generalized) entropy

Entropies.Renyi — Type

Renyi <: Entropy
Renyi(q, base = 2)
Renyi(; q = 1.0, base = 2)

The Rényi[Rényi1960] generalized order-q entropy, used with entropy to compute an entropy with units given by base (typically 2 or MathConstants.e).

Description

Let $p$ be an array of probabilities (summing to 1). Then the Rényi generalized entropy is

\[H_q(p) = \frac{1}{1-q} \log \left(\sum_i p[i]^q\right)\]

and generalizes other known entropies, like e.g. the information entropy ($q = 1$, see [Shannon1948]), the maximum entropy ($q=0$, also known as Hartley entropy), or the correlation entropy ($q = 2$, also known as collision entropy).

If the probability estimator has known alphabet length $L$, then the maximum value of the Rényi entropy is $\log_{base}(L)$, which is the entropy of the uniform distribution with given alphabet length.

Tsallis (generalized) entropy

Entropies.Tsallis — Type

Tsallis <: Entropy
Tsallis(q; k = 1.0, base = 2)
Tsallis(; q = 1.0, k = 1.0, base = 2)

The Tsallis[Tsallis1988] generalized order-q entropy, used with entropy to compute an entropy.

base only applies in the limiting case q == 1, in which the Tsallis entropy reduces to Shannon entropy.

Description

The Tsallis entropy is a generalization of the Boltzmann-Gibbs entropy, with k standing for the Boltzmann constant. It is defined as

\[S_q(p) = \frac{k}{q - 1}\left(1 - \sum_{i} p[i]^q\right)\]

If the probability estimator has known alphabet length $L$, then the maximum value of the Tsallis entropy is $k(L^{1 - q} - 1)/(1 - q)$.

Shannon entropy (convenience)

Entropies.Shannon — Function

Shannon(; base = 2)

The Shannon[Shannon1948] entropy, used with entropy to compute:

\[H(p) = - \sum_i p[i] \log(p[i])\]

with the $log$ at the given base.

Shannon(base) is syntactically equivalent to Renyi(; base).

Curado entropy

Entropies.Curado — Type

Curado <: Entropy
Curado(; b = 1.0)

The Curado entropy (Curado & Nobre, 2004)[Curado2004], used with entropy to compute

\[H_C(p) = \left( \sum_{i=1}^N e^{-b p_i} \right) + e^{-b} - 1,\]

with b ∈ ℛ, b > 0, where the terms outside the sum ensures that $H_C(0) = H_C(1) = 0$.

Stretched exponental entropy

Entropies.StretchedExponential — Type

StretchedExponential <: Entropy
StretchedExponential(; η = 2.0, base = 2)

The stretched exponential, or Anteneodo-Plastino, entropy (Anteneodo & Plastino, 1999[Anteneodo1999]), used with entropy to compute

\[S_{\eta}(p) = \sum_{i = 1}^N \Gamma \left( \dfrac{\eta + 1}{\eta}, - \log_{base}(p_i) \right) - p_i \Gamma \left( \dfrac{\eta + 1}{\eta} \right),\]

where $\eta \geq 0$, $\Gamma(\cdot, \cdot)$ is the upper incomplete Gamma function, and $\Gamma(\cdot) = \Gamma(\cdot, 0)$ is the Gamma function. Reduces to Shannon entropy for η = 1.0.

Normalized entropies

Entropies.entropy_maximum — Function

entropy_maximum(e::Entropy, x, est::ProbabilitiesEstimator) → m::Real

Return the maximum value m of the given entropy type based on the given estimator and the given input x (whose values are not important, but layout and type are).

This function only works if the maximum value is deducable, which is possible only when the estimator has a known total_outcomes.

entropy_maximum(e::Entropy, L::Int) → m::Real

Alternatively, compute the maximum entropy from the total outcomes L directly.

Entropies.entropy_normalized — Function

entropy_normalized([e::Entropy,] x, est::ProbabilitiesEstimator) → h̃ ∈ [0, 1]

Return the normalized entropy of x, i.e., the value of entropy divided by the maximum value for e, according to the given probability estimator. If e is not given, it defaults to Shannon().

Notice that unlike for entropy, here there is no method entropy_normalized(e::Entropy, probs::Probabilities) because there is no way to know the amount of possible events (i.e., the total_outcomes) from probs.

Indirect entropies

Here we list functions which compute Shannon entropies via alternate means, without explicitly computing some probability distributions and then using the Shannon formula.

Nearest neighbors entropy

Entropies.Kraskov — Type

Kraskov <: IndirectEntropy
Kraskov(; k::Int = 1, w::Int = 1, base = 2)

An indirect entropy used in entropy(Kraskov(), x) to estimate the Shannon entropy of x (a multi-dimensional Dataset) to the given base using k-th nearest neighbor searches as in [Kraskov2004].

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

See also: KozachenkoLeonenko.

Entropies.KozachenkoLeonenko — Type

KozachenkoLeonenko <: IndirectEntropy
KozachenkoLeonenko(; k::Int = 1, w::Int = 1, base = 2)

An indirect entropy estimator used in entropy(KozachenkoLeonenko(), x) to estimate the Shannon entropy of x (a multi-dimensional Dataset) to the given base using nearest neighbor searches using the method from Kozachenko & Leonenko[KozachenkoLeonenko1987], as described in Charzyńska and Gambin[Charzyńska2016].

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

In contrast to Kraskov, this estimator uses only the closest neighbor.

Entropies.Zhu — Type

Zhu <: IndirectEntropy
Zhu(k = 1, w = 0, base = MathConstants.e)

The Zhu indirect entropy estimator (Zhu et al., 2015)[Zhu2015] estimates the Shannon entropy of x (a multi-dimensional Dataset) to the given base, by approximating probabilities within hyperrectangles surrounding each point xᵢ ∈ x using using k nearest neighbor searches.

This estimator is an extension to KozachenkoLeonenko.

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

Entropies.ZhuSingh — Type

ZhuSingh <: IndirectEntropy
ZhuSingh(k = 1, w = 0, base = MathConstants.e)

The ZhuSingh indirect entropy estimator (Zhu et al., 2015)[Zhu2015] estimates the Shannon entropy of x (a multi-dimensional Dataset) to the given base.

Like Zhu, this estimator approximates probabilities within hyperrectangles surrounding each point xᵢ ∈ x using using k nearest neighbor searches. However, it also considers the number of neighbors falling on the borders of these hyperrectangles. This estimator is an extension to the entropy estimator in Singh et al. (2003).

w is the Theiler window, which determines if temporal neighbors are excluded during neighbor searches (defaults to 0, meaning that only the point itself is excluded when searching for neighbours).

Order statistics entropy

Entropies.Vasicek — Type

Vasicek <: IndirectEntropy
Vasicek(; m::Int = 1, base = 2)

An indirect entropy estimator used in entropy(Vasicek(), x) to estimate the Shannon entropy of the timeseries x to the given base using the method from Vasicek (1976)[Vasicek1976].

Description

The Vasicek entropy estimator first computes the order statistics $X_{(1)} \leq X_{(2)} \leq \cdots \leq X_{(n)}$ for a random sample of length $n$, i.e. the input timeseries. The Shannon entropy is then estimated as

\[H_V(m) = \dfrac{1}{n} \sum_{i = 1}^n \log \left[ \dfrac{n}{2m} (X_{(i+m)} - X_{(i-m)}) \right]\]

Usage

In practice, choice of m influences how fast the entropy converges to the true value. For small value of m, convergence is slow, so we recommend to scale m according to the time series length n and use m >= n/100 (this is just a heuristic based on the tests written for this package).

Entropies.AlizadehArghami — Type

AlizadehArghami <: IndirectEntropy AlizadehArghami(; m::Int = 1, base = 2)

An indirect entropy estimator used in entropy(Alizadeh(), x) to estimate the Shannon entropy of the timeseries x to the given base using the method from Alizadeh & Arghami (2010)[Alizadeh2010].

Description

The Alizadeh entropy estimator first computes the order statistics $X_{(1)} \leq X_{(2)} \leq \cdots \leq X_{(n)}$ for a random sample of length $n$, i.e. the input timeseries. The entropy for the length-n sample is then estimated as the Vasicek entropy estimate, plus a correction factor

\[H_{A}(m, n) = H_{V}(m, n) + \dfrac{2}{n}\left(m \log_{base}(2) \right).\]

Entropies.Ebrahimi — Type

Ebrahimi <: IndirectEntropy
Ebrahimi(; m::Int = 1, base = 2)

An indirect entropy estimator used in entropy(Ebrahimi(), x) to estimate the Shannon entropy of the timeseries x to the given base using the method from Ebrahimi (1994)[Ebrahimi1994].

Description

The Ebrahimi entropy estimator first computes the order statistics $X_{(1)} \leq X_{(2)} \leq \cdots \leq X_{(n)}$ for a random sample of length $n$, i.e. the input timeseries. The Shannon entropy is then estimated as

\[H_{E}(m) = \dfrac{1}{n} \sum_{i = 1}^n \log \left[ \dfrac{n}{c_i m} (X_{(i+m)} - X_{(i-m)}) \right],\]

where

\[c_i = \begin{cases} 1 + \frac{i - 1}{m}, & 1 \geq i \geq m \ 2, & m + 1 \geq i \geq n - m \ 1 + \frac{n - i}{m} & n - m + 1 \geq i \geq n \end{cases}.\]

Entropies.Correa — Type

Correa <: IndirectEntropy
Correa(; m::Int = 1, base = 2)

An indirect entropy estimator used in entropy(Correa(), x) to estimate the Shannon entropy of the timeseries x to the given base using the method from Correa (1995)[Correa1995].

Description

The Correa entropy estimator first computes the order statistics like Vasicek, ensuring that edge points are included, then estimates entropy as

\[H_C(m, n) = \dfrac{1}{n} \sum_{i = 1}^n \log \left[ \dfrac{ \sum_{j=i-m}^{i+m}(X_{(j)} - \bar{X}_{(i)})(j - i)}{n \sum_{j=i-m}^{i+m} (X_{(j)} - \bar{X}_{(i)})^2} \right],\]

where

\[\bar{X}_{(i)} = \dfrac{1}{2m + 1} \sum_{j = i - m}^{i + m} X_{(j)}.\]

Convenience functions

In this subsection we expand documentation strings of "entropy names" that are used commonly in the literature, such as "permutation entropy". As we made clear in API & terminology, these are just the existing Shannon entropy with a particularly chosen probability estimator. We have only defined convenience functions for the most used names, and arbitrary more specialized convenience functions can be easily defined in a couple lines of code.

Entropies.entropy_permutation — Function

entropy_permutation(x; τ = 1, m = 3, base = 2)

Compute the permutation entropy of x of order m with delay/lag τ. This function is just a convenience call to:

est = SymbolicPermutation(; m, τ)
entropy(Shannon(base), x, est)

See SymbolicPermutation for more info. Similarly, one can use SymbolicWeightedPermutation or SymbolicAmplitudeAwarePermutation for the weighted/amplitude-aware versions.

Entropies.entropy_spatial_permutation — Function

entropy_spatial_permutation(x, stencil, periodic = true; kwargs...)

Compute the spatial permutation entropy of x given the stencil. Here x must be a matrix or higher dimensional Array containing spatial data. This function is just a convenience call to:

est = SpatialSymbolicPermutation(stencil, x, periodic)
entropy(Renyi(;kwargs...), x, est)

See SpatialSymbolicPermutation for more info, or how to encode stencils.

Entropies.entropy_wavelet — Function

entropy_wavelet(x; wavelet = Wavelets.WT.Daubechies{12}(), base = 2)

Compute the wavelet entropy. This function is just a convenience call to:

est = WaveletOverlap(wavelet)
entropy(Shannon(base), x, est)

See WaveletOverlap for more info.

Entropies.entropy_dispersion — Function

entropy_dispersion(x; base = 2, kwargs...)

Compute the dispersion entropy. This function is just a convenience call to:

est = Dispersion(kwargs...)
entropy(Shannon(base), x, est)

See Dispersion for more info.

Rényi1960A. Rényi, Proceedings of the fourth Berkeley Symposium on Mathematics, Statistics and Probability, pp 547 (1960)

Shannon1948C. E. Shannon, Bell Systems Technical Journal 27, pp 379 (1948)

Tsallis1988Tsallis, C. (1988). Possible generalization of Boltzmann-Gibbs statistics. Journal of statistical physics, 52(1), 479-487.

Shannon1948C. E. Shannon, Bell Systems Technical Journal 27, pp 379 (1948)

Curado2004Curado, E. M., & Nobre, F. D. (2004). On the stability of analytic entropic forms. Physica A: Statistical Mechanics and its Applications, 335(1-2), 94-106.

Anteneodo1999Anteneodo, C., & Plastino, A. R. (1999). Maximum entropy approach to stretched exponential probability distributions. Journal of Physics A: Mathematical and General, 32(7), 1089.

Kraskov2004Kraskov, A., Stögbauer, H., & Grassberger, P. (2004). Estimating mutual information. Physical review E, 69(6), 066138.

Charzyńska2016Charzyńska, A., & Gambin, A. (2016). Improvement of the k-NN entropy estimator with applications in systems biology. Entropy, 18(1), 13.

KozachenkoLeonenko1987Kozachenko, L. F., & Leonenko, N. N. (1987). Sample estimate of the entropy of a random vector. Problemy Peredachi Informatsii, 23(2), 9-16.

Zhu2015Zhu, J., Bellanger, J. J., Shu, H., & Le Bouquin Jeannès, R. (2015). Contribution to transfer entropy estimation via the k-nearest-neighbors approach. Entropy, 17(6), 4173-4201.

Zhu2015Zhu, J., Bellanger, J. J., Shu, H., & Le Bouquin Jeannès, R. (2015). Contribution to transfer entropy estimation via the k-nearest-neighbors approach. Entropy, 17(6), 4173-4201.

Singh2003Singh, H., Misra, N., Hnizdo, V., Fedorowicz, A., & Demchuk, E. (2003). Nearest neighbor estimates of entropy. American journal of mathematical and management sciences, 23(3-4), 301-321.

Vasicek1976Vasicek, O. (1976). A test for normality based on sample entropy. Journal of the Royal Statistical Society: Series B (Methodological), 38(1), 54-59.

Alizadeh2010Alizadeh, N. H., & Arghami, N. R. (2010). A new estimator of entropy. Journal of the Iranian Statistical Society (JIRSS).

Ebrahimi1994Ebrahimi, N., Pflughoeft, K., & Soofi, E. S. (1994). Two measures of sample entropy. Statistics & Probability Letters, 20(3), 225-234.

Correa1995Correa, J. C. (1995). A new estimator of entropy. Communications in Statistics-Theory and Methods, 24(10), 2439-2449.