Generalized entropy

Entropies.genentropy — Function

genentropy(α::Real, p::AbstractArray; base = Base.MathConstants.e)

Compute the entropy, to the given base, of an array of probabilities p, assuming that p is sum-normalized.

Description

Let $p$ be an array of probabilities (summing to 1). Then the Rényi entropy is

\[H_\alpha(p) = \frac{1}{1-\alpha} \log \left(\sum_i p[i]^\alpha\right)\]

and generalizes other known entropies, like e.g. the information entropy ($\alpha = 1$, see [Shannon1948]), the maximum entropy ($\alpha=0$, also known as Hartley entropy), or the correlation entropy ($\alpha = 2$, also known as collision entropy).

Example

using Entropies
p = rand(5000)
p = p ./ sum(p) # normalizing to 1 ensures we have a probability distribution

# Estimate order-1 generalized entropy to base 2 of the distribution
Entropies.genentropy(1, ps, base = 2)

using DelayEmbeddings, Entropies

# Some random data, and its corresponding sum-normalized histogram (which sums to 1, so is a 
# probability distribution)
D = Dataset(rand(1:3, 10000, 3))
ps = Entropies.non0hist(D)

# Estimate order-1 generalized entropy to base 2 of the distribution
Entropies.genentropy(1, ps, base = 2)

See also: non0hist.

genentropy(α::Real, x::Dataset, est::VisitationFrequency)

Compute the α order generalized (Rényi) entropy[Rényi1960] of a multivariate dataset x.

Description

First, the state space defined by x is partitioned into rectangular boxes according to the binning instructions given by est.binning. Then, a histogram of visitations to each of those boxes is obtained, which is then sum-normalized to obtain a probability distribution, using probabilities. The generalized entropy to base est.b is then computed over that box visitation distribution using genentropy(::Real, ::AbstractArray).

Example

using DelayEmbeddings, Entropies

# Estimator specification. Split each coordinate axis in five equal segments, and 
# use logarithms to base 2.
est = VisitationFrequency(RectangularBinning(5), b = 2) 

# Estimate order-1 generalized entropy of some random data
D = Dataset(rand(1:3, 20000, 3))

Entropies.genentropy(1, D, est)

See also: VisitationFrequency.

Rényi1960A. Rényi, Proceedings of the fourth Berkeley Symposium on Mathematics, Statistics and Probability, pp 547 (1960)

Shannon1948C. E. Shannon, Bell Systems Technical Journal 27, pp 379 (1948)