Probabilities

Probabilities API

The probabilities API is defined by

ProbabilitiesEstimator

probabilities

probabilities_and_outcomes

Entropies.ProbabilitiesEstimator — Type

ProbabilitiesEstimator

The supertype for all probabilities estimators.

In Entropies.jl, probability distributions are estimated from data by defining a set of possible outcomes $\Omega = \{\omega_1, \omega_2, \ldots, \omega_L \}$, and assigning to each outcome $\omega_i$ a probability $p(\omega_i)$, such that $\sum_{i=1}^N p(\omega_i) = 1$. It is the role a ProbabilitiesEstimator to

Define $\Omega$, the "outcome space", which is the set of all possible outcomes over which probabilities are estimated. The cardinality of this set can be obtained using total_outcomes.

Define how probabilities $p_i = p(\omega_i)` are assigned to outcomes$\omega_i``.

In practice, probability estimation is done by calling probabilities with some input data and one of the following probabilities estimators. The result is a Probabilities p (Vector-like), where each element p[i] is the probability of the outcome ω[i]. Use probabilities_and_outcomes if you need both the probabilities and the outcomes and outcome_space to obtain $\Omega$. The element type of $\Omega$ varies between estimators, but it is guranteed to be hashable. This allows for conveniently tracking the probability of a specific event across experimental realizations, by using the outcome as a dictionary key and the probability as the value for that key (or, alternatively, the key remains the outcome and one has a vector of probabilities, one for each experimental realization).

We have made the design decision that all probabilities estimators have a well defined outcome space when instantiated. For some estimators this means that the input data x must be provided both when instantiating the estimator, but also when computing the probabilities.

All currently implemented probability estimators are:

CountOccurrences.

ValueHistogram.

TransferOperator.

Dispersion.

SpatialDispersion.

WaveletOverlap.

PowerSpectrum.

SymbolicPermutation.

SymbolicWeightedPermutation.

SymbolicAmplitudeAwarePermutation.

SpatialSymbolicPermutation.

NaiveKernel.

Entropies.probabilities — Function

probabilities(est::ProbabilitiesEstimator, x::Array_or_Dataset) → p::Probabilities

Compute a probability distribution over the set of possible outcomes defined by the probabilities estimator est, given input data x. To obtain the outcomes use outcomes.

The returned probabilities p may or may not be ordered, and may or may not contain 0s; see the documentation of the individual estimators for more. Configuration options are always given as arguments to the chosen estimator. x is typically an Array or a Dataset; see Input data for Entropies.jl.

probabilities(x::Array_or_Dataset) → p::Probabilities

Estimate probabilities by directly counting the elements of x, assuming that Ω = sort(unique(x)), i.e. that the outcome space is the unique elements of x. This is mostly useful when x contains categorical or integer data.

See also: Probabilities, ProbabilitiesEstimator.

Entropies.probabilities! — Function

probabilities!(s, args...)

Similar to probabilities(args...), but allows pre-allocation of temporarily used containers s.

Only works for certain estimators. See for example SymbolicPermutation.

Entropies.Probabilities — Type

Probabilities <: AbstractVector
Probabilities(x) → p

Probabilities is a simple wrapper around AbstractVector that ensures its values sum to 1, so that p can be interpreted as probability distribution.

Entropies.probabilities_and_outcomes — Function

probabilities_and_outcomes(est, x)

Return probs, outs, where probs = probabilities(x, est) and outs[i] is the outcome with probability probs[i]. The element type of outs depends on the estimator. outs is a subset of the outcome_space of est.

See also outcomes, total_outcomes.

Entropies.outcomes — Function

outcomes(est::ProbabilitiesEstimator, x)

Return all (unique) outcomes contained in x according to the given estimator. Equivalent with probabilities_and_outcomes(x, est)[2], but for some estimators it may be explicitly extended for better performance.

Entropies.outcome_space — Function

outcome_space(est::ProbabilitiesEstimator) → Ω

Return a container containing all possible outcomes of est.

Entropies.total_outcomes — Function

total_outcomes(est::ProbabilitiesEstimator)

Return the length (cardinality) of the outcome space $\Omega$ of est.

Entropies.missing_outcomes — Function

missing_outcomes(est::ProbabilitiesEstimator, x) → n_missing::Int

Estimate a probability distribution for x using the given estimator, then count the number of missing (i.e. zero-probability) outcomes.

See also: MissingDispersionPatterns.

Overview

Any of the following estimators can be used with probabilities.

Estimator	Principle	Input data

CountOccurrences	Frequencies	Vector, Dataset

ValueHistogram	Binning (histogram)	Vector, Dataset

TransferOperator	Binning (transfer operator)	Vector, Dataset

NaiveKernel	Kernel density estimation	Dataset

SymbolicPermutation	Ordinal patterns	Vector

SymbolicWeightedPermutation	Ordinal patterns	Vector

SymbolicAmplitudeAwarePermutation	Ordinal patterns	Vector

Dispersion	Dispersion patterns	Vector

Diversity	Cosine similarity	Vector

WaveletOverlap	Wavelet transform	Vector

PowerSpectrum	Fourier spectra	Vector, Dataset

Count occurrences (counting)

Entropies.CountOccurrences — Type

CountOccurrences(x)

A probabilities/entropy estimator based on straight-forward counting of distinct elements in a univariate time series or multivariate dataset. This is the same as giving no estimator to probabilities.

Outcome space

The outcome space is the unique sorted values of the input. Hence, input x is needed for a well-defined outcome space.

Visitation frequency (histograms)

Entropies.ValueHistogram — Type

ValueHistogram(x, b::RectangularBinning) <: ProbabilitiesEstimator
ValueHistogram(b::FixedRectangularBinning) <: ProbabilitiesEstimator

A probability estimator based on binning the values of the data as dictated by the binning scheme b and formally computing their histogram, i.e., the frequencies of points in the bins. An alias to this is VisitationFrequency. Available binnings are:

RectangularBinning

FixedRectangularBinning

Notice that if not using the fixed binning, x (the input data) must also be given to the estimator, as it is not possible to deduce histogram size only from the binning.

The ValueHistogram estimator has a linearithmic time complexity (n log(n) for n = length(x)) and a linear space complexity (l for l = dimension(x)). This allows computation of probabilities (histograms) of high-dimensional datasets and with small box sizes ε without memory overflow and with maximum performance. For performance reasons, the probabilities returned never contain 0s and are arbitrarily ordered.

ValueHistogram(x, ϵ::Union{Real,Vector})

A convenience method that accepts same input as RectangularBinning and initializes this binning directly.

Outcomes

The outcome space for ValueHistogram is the unique bins constructed from b. Each bin is identified by its left (lowest-value) corner. The bins are in data units, not integer (cartesian indices units), and are returned as SVectors.

See also: RectangularBinning.

Entropies.RectangularBinning — Type

RectangularBinning(ϵ) <: AbstractBinning

Rectangular box partition of state space using the scheme ϵ, deducing the coordinates of the grid axis minima from the input data. Generally it is preferred to use FixedRectangularBinning instead, as it has a well defined outcome space without knowledge of input data.

Binning instructions are deduced from the type of ϵ as follows:

ϵ::Int divides each coordinate axis into ϵ equal-length intervals that cover all data.

ϵ::Float64 divides each coordinate axis into intervals of fixed size ϵ, starting from the axis minima until the data is completely covered by boxes.

ϵ::Vector{Int} divides the i-th coordinate axis into ϵ[i] equal-length intervals that cover all data.

ϵ::Vector{Float64} divides the i-th coordinate axis into intervals of fixed size ϵ[i], starting from the axis minima until the data is completely covered by boxes.

Entropies.FixedRectangularBinning — Type

FixedRectangularBinning <: AbstractBinning
FixedRectangularBinning(ϵmin::NTuple, ϵmax::NTuple, N::Int)

Rectangular box partition of state space where the extent of the grid is explicitly specified by ϵmin and emax, and along each dimension, the grid is subdivided into N subintervals. Points falling outside the partition do not attribute to probabilities. This binning type leads to a well-defined outcome space without knowledge of input, see ValueHistogram.

ϵmin/emax must be NTuple{D, <:Real} for input of D-dimensional data.

FixedRectangularBinning(ϵmin::Real, ϵmax::Real, N::Int, D::Int = 1)

This is a convenience method where each dimension of the binning has the same extent and the input data are D dimensional, which defaults to 1 (timeseries).

Permutation (symbolic)

Entropies.SymbolicPermutation — Type

SymbolicPermutation <: ProbabilitiesEstimator
SymbolicPermutation(; m = 3, τ = 1, lt::Function = Entropies.isless_rand)

A probabilities estimator based on ordinal permutation patterns.

The quantity computed depends on the input data:

Univariate data. If applied to a univariate time series, then the time series is first embedded using embedding delay τ and dimension m, resulting in embedding vectors $\{ \bf{x}_i \}_{i=1}^{N-(m-1)\tau}$. Then, for each $\bf{x}_i$, we find its permutation pattern $\pi_{i}$, which we internally encode a an integer $s_i \in \mathbb{N}^+$ for efficient computation (integer symbols are obtained by using encode with OrdinalPatternEncoding). Probabilities are then estimated as naive frequencies over the encoded permutation symbols $\{ s_i \}_{i=1}^{N-(m-1)\tau}$ by using CountOccurrences. The resulting probabilities can be used to compute permutation entropy (PE; Bandt & Pompe, 2002[BandtPompe2002]).

Multivariate data. If applied to a an D-dimensional Dataset, then it is assumed that the input data represents $N$ observations of a multivariate system $\{ \bf{x}_i \}_{i=1}^N$, and no embedding is constructed. For each $\bf{x}_i \in \mathbb{R}^D$, we direct find its permutation pattern $\pi_{i}$ and encode it as $s_i \in \mathbb{N}^+$ (i.e. est.τ and est.m are ignored, and we set m = D instead). Finally, probabilities are estimated as relative frequencies of occurrences of the encoded permutation symbols. The resulting probabilities can be used to compute multivariate permutation entropy (MvPE; He et al., 2016[He2016]), but here we don't perform any subdivision of the permutation patterns (see Figure 3 in He et al., 2016).

Outcome space

The outcome space Ω for SymbolicPermutation is the set of length-m ordinal patterns (i.e. permutations) that can be formed by the integers 1, 2, …, m, ordered lexicographically. There are factorial(m) such patterns.

In-place symbolization

SymbolicPermutation also implements the in-place entropy! and probabilities!. The length of the pre-allocated symbol vector must match the length of the embedding: N - (m-1)τ for univariate time series, and M for length-M Datasets), i.e.

using DelayEmbeddings, Entropies
m, τ, N = 2, 1, 100
est = SymbolicPermutation(; m, τ)

# For a time series
x_ts = rand(N)
πs_ts = zeros(Int, N - (m - 1)*τ)
p = probabilities!(πs_ts, est, x_ts)
h = entropy!(πs_ts, Renyi(), est, x_ts)

# For a pre-discretized `Dataset`
x_symb = outcomes(x_ts, OrdinalPatternEncoding(m = 2, τ = 1))
x_d = genembed(x_symb, (0, -1, -2))
πs_d = zeros(Int, length(x_d))
p = probabilities!(πs_d, est, x_d)
h = entropy!(πs_d, Renyi(), est, x_d)

See SymbolicWeightedPermutation and SymbolicAmplitudeAwarePermutation for estimators that not only consider ordinal (sorting) patterns, but also incorporate information about within-state-vector amplitudes.

Handling equal values in ordinal patterns

In Bandt & Pompe (2002), equal values are ordered after their order of appearance, but this can lead to erroneous temporal correlations, especially for data with low-amplitude resolution [Zunino2017]. Here, by default, if two values are equal, then one of the is randomly assigned as "the largest", using lt = Entropies.isless_rand. To get the behaviour from Bandt and Pompe (2002), use lt = Base.isless).

Entropies.SymbolicWeightedPermutation — Type

SymbolicWeightedPermutation <: ProbabilitiesEstimator
SymbolicWeightedPermutation(; τ = 1, m = 3, lt = Entropies.isless_rand)

A variant of SymbolicPermutation that also incorporates amplitude information, based on the weighted permutation entropy (Fadlallah et al., 2013).

Outcome space

Like for SymbolicPermutation, the outcome space Ω for SymbolicWeightedPermutation is the lexiographically ordered set of length-m ordinal patterns (i.e. permutations) that can be formed by the integers 1, 2, …, m. There are factorial(m) such patterns.

Description

Probabilities are computed as

\[p(\pi_i^{m, \tau}) = \dfrac{\sum_{k=1}^N \mathbf{1}_{u:S(u) = s_i} \left( \mathbf{x}_k^{m, \tau} \right) \, w_k}{\sum_{k=1}^N \mathbf{1}_{u:S(u) \in \Pi} \left( \mathbf{x}_k^{m, \tau} \right) \,w_k} = \dfrac{\sum_{k=1}^N \mathbf{1}_{u:S(u) = s_i} \left( \mathbf{x}_k^{m, \tau} \right) \, w_k}{\sum_{k=1}^N w_k},\]

where weights are computed based on the variance of the state vectors as

\[w_j = \dfrac{1}{m}\sum_{k=1}^m (x_{j+(k+1)\tau} - \mathbf{\hat{x}}_j^{m, \tau})^2,\]

and $\mathbf{x}_i$ is the aritmetic mean of state vector:

\[\mathbf{\hat{x}}_j^{m, \tau} = \frac{1}{m} \sum_{k=1}^m x_{j + (k+1)\tau}.\]

The weighted permutation entropy is equivalent to regular permutation entropy when weights are positive and identical ($w_j = \beta \,\,\, \forall \,\,\, j \leq N$ and $\beta > 0)$.

See SymbolicPermutation for an estimator that only incorporates ordinal/sorting information and disregards amplitudes, and SymbolicAmplitudeAwarePermutation for another estimator that incorporates amplitude information.

An implementation note

Note: in equation 7, section III, of the original paper, the authors write

\[w_j = \dfrac{1}{m}\sum_{k=1}^m (x_{j-(k-1)\tau} - \mathbf{\hat{x}}_j^{m, \tau})^2.\]

But given the formula they give for the arithmetic mean, this is not the variance of $\mathbf{x}_i$, because the indices are mixed: $x_{j+(k-1)\tau}$ in the weights formula, vs. $x_{j+(k+1)\tau}$ in the arithmetic mean formula. This seems to imply that amplitude information about previous delay vectors are mixed with mean amplitude information about current vectors. The authors also mix the terms "vector" and "neighboring vector" (but uses the same notation for both), making it hard to interpret whether the sign switch is a typo or intended. Here, we use the notation above, which actually computes the variance for $\mathbf{x}_i$.

Entropies.SymbolicAmplitudeAwarePermutation — Type

SymbolicAmplitudeAwarePermutation <: ProbabilitiesEstimator
SymbolicAmplitudeAwarePermutation(; τ = 1, m = 3, A = 0.5, lt = Entropies.isless_rand)

A variant of SymbolicPermutation that also incorporates amplitude information, based on the amplitude-aware permutation entropy (Azami & Escudero, 2016).

Outcome space

Like for SymbolicPermutation, the outcome space Ω for SymbolicAmplitudeAwarePermutation is the lexiographically ordered set of length-m ordinal patterns (i.e. permutations) that can be formed by the integers 1, 2, …, m. There are factorial(m) such patterns.

Description

Probabilities are computed as

\[p(\pi_i^{m, \tau}) = \dfrac{\sum_{k=1}^N \mathbf{1}_{u:S(u) = s_i} \left( \mathbf{x}_k^{m, \tau} \right) \, a_k}{\sum_{k=1}^N \mathbf{1}_{u:S(u) \in \Pi} \left( \mathbf{x}_k^{m, \tau} \right) \,a_k} = \dfrac{\sum_{k=1}^N \mathbf{1}_{u:S(u) = s_i} \left( \mathbf{x}_k^{m, \tau} \right) \, a_k}{\sum_{k=1}^N a_k}.\]

The weights encoding amplitude information about state vector $\mathbf{x}_i = (x_1^i, x_2^i, \ldots, x_m^i)$ are

\[a_i = \dfrac{A}{m} \sum_{k=1}^m |x_k^i | + \dfrac{1-A}{d-1} \sum_{k=2}^d |x_{k}^i - x_{k-1}^i|,\]

with $0 \leq A \leq 1$. When $A=0$ , only internal differences between the elements of $\mathbf{x}_i$ are weighted. Only mean amplitude of the state vector elements are weighted when $A=1$. With, $0<A<1$, a combined weighting is used.

See SymbolicPermutation for an estimator that only incorporates ordinal/sorting information and disregards amplitudes, and SymbolicWeightedPermutation for another estimator that incorporates amplitude information.

Entropies.SpatialSymbolicPermutation — Type

SpatialSymbolicPermutation <: ProbabilitiesEstimator
SpatialSymbolicPermutation(stencil, x; periodic = true)

A symbolic, permutation-based probabilities estimator for spatiotemporal systems that generalises SymbolicPermutation to high-dimensional arrays.

SpatialSymbolicPermutation is based on the 2D and 3D spatiotemporal permutation entropy estimators by by Ribeiro et al. (2012)[Ribeiro2012] and Schlemmer et al. (2018)[Schlemmer2018]), respectively, but is here implemented as a pure probabilities probabilities estimator that is generalized for N-dimensional input data x, with arbitrary neighborhood regions (stencils) and periodic boundary conditions.

In combination with entropy and entropy_normalized, this probabilities estimator can be used to compute (normalized) generalized spatiotemporal permutation Entropy of any type.

Arguments

stencil. Defines what local area (hyperrectangle), or which points within this area, to include around each hypervoxel (i.e. pixel in 2D). The examples below demonstrate different ways of specifying stencils. For details, see SpatialSymbolicPermutation.

x::AbstractArray. The input data. Must be provided because we need to know its size for optimization and bound checking.

Keyword arguments

periodic::Bool. If periodic == true, then the stencil should wrap around at the end of the array. If periodic = false, then pixels whose stencil exceeds the array bounds are skipped.

Stencils

Stencils are passed in one of the following three ways:

As vectors of CartesianIndex which encode the pixels to include in the stencil, with respect to the current pixel, or integer arrays of the same dimensionality as the data. For example stencil = CartesianIndex.([(0,0), (0,1), (1,1), (1,0)]). Don't forget to include the zero offset index if you want to include the point itself, which is almost always the case. Here the stencil creates a 2x2 square extending to the bottom and right of the pixel (directions here correspond to the way Julia prints matrices by default). When passing a stencil as a vector of CartesianIndex, m = length(stencil).

As a D-dimensional array (where D matches the dimensionality of the input data) containing 0s and 1s, where if stencil[index] == 1, the corresponding pixel is included, and if stencil[index] == 0, it is not included. To generate the same estimator as in 1., use stencil = [1 1; 1 1]. When passing a stencil as a D-dimensional array, m = sum(stencil)

As a Tuple containing two Tuples, both of length D, for D-dimensional data. The first tuple specifies the extent of the stencil, where extent[i] dictates the number of pixels to be included along the ith axis and lag[i] the separation of pixels along the same axis. This method can only generate (hyper)rectangular stencils. To create the same estimator as in the previous examples, use here stencil = ((2, 2), (1, 1)). When passing a stencil using extent and lag, m = prod(extent)!.

Example: spatiotemporal entropy for time series

Usage is simple. First, define a SpatialSymbolicPermutation estimator by specifying a stencil and giving some input data (a matrix with the same dimensions as the data as you're going to analyse). Then simply call entropy with the estimator.

using Entropies
x = rand(50, 50) # first "time slice" of a spatial system evolution
stencil = [1 1; 0 1] # or one of the other ways of specifying stencils
est = SpatialSymbolicPermutation(stencil, x)
h = entropy(est, x)

To apply this to timeseries of spatial data, simply loop over the call, e.g.:

data = [rand(50, 50) for i in 1:50]
est = SpatialSymbolicPermutation(stencil, first(data))
h_vs_t = [entropy(est, d) for d in data]

Computing generalized spatiotemporal permutation entropy is trivial, e.g. with Renyi:

x = reshape(repeat(1:5, 500) .+ 0.1*rand(500*5), 50, 50)
est = SpatialSymbolicPermutation(stencil, x)
entropy(Renyi(q = 2), est, x)

Outcome space

The outcome space Ω for SpatialSymbolicPermutation is the set of length-m ordinal patterns (i.e. permutations) that can be formed by the integers 1, 2, …, m, ordered lexicographically. There are factorial(m) such patterns. Here, m refers to the number of points included by stencil.

Dispersion (symbolic)

Entropies.Dispersion — Type

Dispersion(; c = 5, m = 2, τ = 1, check_unique = true)

A probability estimator based on dispersion patterns, originally used by Rostaghi & Azami, 2016[Rostaghi2016] to compute the "dispersion entropy", which characterizes the complexity and irregularity of a time series.

Recommended parameter values[Li2018] are m ∈ [2, 3], τ = 1 for the embedding, and c ∈ [3, 4, …, 8] categories for the Gaussian symbol mapping.

Description

Assume we have a univariate time series $X = \{x_i\}_{i=1}^N$. First, this time series is discretized a "Gaussian encoding", which uses the normal cumulative distribution function (CDF) to encode a timeseries $x$ as integers like so:

Each $x_i$ is mapped to a new real number $y_i \in [0, 1]$ by using the normal cumulative distribution function (CDF), $x_i \to y_i : y_i = \dfrac{1}{ \sigma \sqrt{2 \pi}} \int_{-\infty}^{x_i} e^{(-(x_i - \mu)^2)/(2 \sigma^2)} dx$, where $\mu$ and $\sigma$ are the empirical mean and standard deviation of $X$. Other choices of CDFs are also possible, but currently only Gaussian is implemented.

Next, each $y_i$ is linearly mapped to an integer $z_i \in [1, 2, \ldots, c]$ using the map $y_i \to z_i : z_i = \text{floor}(y_i ./ (1 / c)) + 1$ (note: this mapping differs slightly from the linear mapping in the original paper). This procedure subdivides the interval $[0, 1]$ into $c$ different subintervals that form a covering of $[0, 1]$, and assigns each $y_i$ to one of these subintervals. The original time series $X$ is thus transformed to a symbol time series $S = \{ s_i \}_{i=1}^N$, where $s_i \in [1, 2, \ldots, c]$.

Next, the symbol time series $S$ is embedded into an $m$-dimensional time series, using an embedding lag of $\tau = 1$, which yields a total of $N - (m - 1)\tau$ points, or "dispersion patterns". Because each $z_i$ can take on $c$ different values, and each embedding point has $m$ values, there are $c^m$ possible dispersion patterns. This number is used for normalization when computing dispersion entropy.

Computing dispersion probabilities and entropy

A probability distribution $P = \{p_i \}_{i=1}^{c^m}$, where $\sum_i^{c^m} p_i = 1$, can then be estimated by counting and sum-normalising the distribution of dispersion patterns among the embedding vectors. Note that dispersion patterns that are not present are not counted. Therefore, you'll always get non-zero probabilities using the Dispersion probability estimator.

Outcome space

The outcome space for Dispersion is the unique delay vectors whose elements are the the symbols (integers) encoded by the Gaussian CDF. Hence, the outcome space is all m-dimensional delay vectors whose elements are all possible values in 1:c. There are c ^ m such vectors.

Data requirements and parameters

The input must have more than one unique element for the Gaussian mapping to be well-defined. Li et al. (2018) recommends that x has at least 1000 data points.

If check_unique == true (default), then it is checked that the input has more than one unique value. If check_unique == false and the input only has one unique element, then a InexactError is thrown when trying to compute probabilities.

Why 'dispersion patterns'?

Each embedding vector is called a "dispersion pattern". Why? Let's consider the case when $m = 5$ and $c = 3$, and use some very imprecise terminology for illustration:

When $c = 3$, values clustering far below mean are in one group, values clustered around the mean are in one group, and values clustering far above the mean are in a third group. Then the embedding vector $[2, 2, 2, 2, 2]$ consists of values that are relatively close together (close to the mean), so it represents a set of numbers that are not very spread out (less dispersed). The embedding vector $[1, 1, 2, 3, 3]$, however, represents numbers that are much more spread out (more dispersed), because the categories representing "outliers" both above and below the mean are represented, not only values close to the mean.

Entropies.SpatialDispersion — Type

SpatialDispersion <: ProbabilitiesEstimator
SpatialDispersion(stencil, x::AbstractArray;
    periodic = true,
    c = 5,
    skip_encoding = false,
    L = nothing,
)

A dispersion-based probabilities estimator that generalises Dispersion for input data that are high-dimensional arrays.

SpatialDispersion is based on Azami et al. (2019)[Azami2019]'s 2D square dispersion (Shannon) entropy estimator, but is here implemented as a pure probabilities probabilities estimator that is generalized for N-dimensional input data x, with arbitrary neighborhood regions (stencils) and (optionally) periodic boundary conditions.

In combination with entropy and entropy_normalized, this probabilities estimator can be used to compute (normalized) generalized spatiotemporal dispersion Entropy of any type.

Arguments

stencil. Defines what local area (hyperrectangle), or which points within this area, to include around each hypervoxel (i.e. pixel in 2D). The examples below demonstrate different ways of specifying stencils. For details, see SpatialSymbolicPermutation. See SpatialSymbolicPermutation for more information about stencils.

x::AbstractArray. The input data. Must be provided because we need to know its size for optimization and bound checking.

Keyword arguments

periodic::Bool. If periodic == true, then the stencil should wrap around at the end of the array. If periodic = false, then pixels whose stencil exceeds the array bounds are skipped.

c::Int. Determines how many discrete categories to use for the Gaussian encoding.

skip_encoding. If skip_encoding == true, encoding is ignored, and dispersion patterns are computed directly from x, under the assumption that L is the alphabet length for x (useful for categorical or integer data). Thus, if skip_encoding == true, then L must also be specified. This is useful for categorical or integer-valued data.

L. If L == nothing (default), then the number of total outcomes is inferred from stencil and encoding. If L is set to an integer, then the data is considered pre-encoded and the number of total outcomes is set to L.

Outcome space

The outcome space for SpatialDispersion is the unique delay vectors whose elements are the the symbols (integers) encoded by the Gaussian CDF. Hence, the outcome space is all m-dimensional delay vectors whose elements are all possible values in 1:c. There are c ^ m such vectors.

Description

Estimating probabilities/entropies from higher-dimensional data is conceptually simple.

Discretize each value (hypervoxel) in x relative to all other values xᵢ ∈ x using the provided encoding scheme.

Use stencil to extract relevant (discretized) points around each hypervoxel.

Construct a symbol these points.

Take the sum-normalized histogram of the symbol as a probability distribution.

Optionally, compute entropy or entropy_normalized from this probability distribution.

Usage

Here's how to compute spatial dispersion entropy using the three different ways of specifying stencils.

x = rand(50, 50) # first "time slice" of a spatial system evolution

# Cartesian stencil
stencil_cartesian = CartesianIndex.([(0,0), (1,0), (1,1), (0,1)])
est = SpatialDispersion(stencil_cartesian, x)
entropy_normalized(est, x)

# Extent/lag stencil
extent = (2, 2); lag = (1, 1); stencil_ext_lag = (extent, lag)
est = SpatialDispersion(stencil_ext_lag, x)
entropy_normalized(est, x)

# Matrix stencil
stencil_matrix = [1 1; 1 1]
est = SpatialDispersion(stencil_matrix, x)
entropy_normalized(est, x)

To apply this to timeseries of spatial data, simply loop over the call (broadcast), e.g.:

imgs = [rand(50, 50) for i = 1:100]; # one image per second over 100 seconds
stencil = ((2, 2), (1, 1)) # a 2x2 stencil (i.e. dispersion patterns of length 4)
est = SpatialDispersion(stencil, first(imgs))
h_vs_t = entropy_normalized.(Ref(est), imgs)

Computing generalized spatiotemporal dispersion entropy is trivial, e.g. with Renyi:

x = reshape(repeat(1:5, 500) .+ 0.1*rand(500*5), 50, 50)
est = SpatialDispersion(stencil, x)
entropy(Renyi(q = 2), est, x)

See also: SpatialSymbolicPermutation, GaussianCDFEncoding, symbolize.

Transfer operator (binning)

Entropies.TransferOperator — Type

TransferOperator <: <ProbabilitiesEstimator
TransferOperator(b::RectangularBinning)

A probability estimator based on binning data into rectangular boxes dictated by the given binning scheme b, then approximating the transfer (Perron-Frobenius) operator over the bins, then taking the invariant measure associated with that transfer operator as the bin probabilities. Assumes that the input data are sequential (time-ordered).

This implementation follows the grid estimator approach in Diego et al. (2019)[Diego2019].

Outcome space

The outcome space for TransferOperator is the set of unique bins constructed from b. Bins are identified by their left (lowest-value) corners, are given in data units, and are returned as SVectors.

Bin ordering

Bins returned by probabilities_and_outcomes are ordered according to first appearance (i.e. the first time the input (multivariate) timeseries visits the bin). Thus, if

b = RectangularBinning(4)
est = TransferOperator(b)
probs, outcomes = probabilities_and_outcomes(x, est) # x is some timeseries

then probs[i] is the invariant measure (probability) of the bin outcomes[i], which is the i-th bin visited by the timeseries with nonzero measure.

Description

The transfer operator $P^{N}$is computed as an N-by-N matrix of transition probabilities between the states defined by the partition elements, where N is the number of boxes in the partition that is visited by the orbit/points.

If $\{x_t^{(D)} \}_{n=1}^L$ are the $L$ different $D$-dimensional points over which the transfer operator is approximated, $\{ C_{k=1}^N \}$ are the $N$ different partition elements (as dictated by ϵ) that gets visited by the points, and $\phi(x_t) = x_{t+1}$, then

\[P_{ij} = \dfrac {\#\{ x_n | \phi(x_n) \in C_j \cap x_n \in C_i \}} {\#\{ x_m | x_m \in C_i \}},\]

where $\#$ denotes the cardinal. The element $P_{ij}$ thus indicates how many points that are initially in box $C_i$ end up in box $C_j$ when the points in $C_i$ are projected one step forward in time. Thus, the row $P_{ik}^N$ where $k \in \{1, 2, \ldots, N \}$ gives the probability of jumping from the state defined by box $C_i$ to any of the other $N$ states. It follows that $\sum_{k=1}^{N} P_{ik} = 1$ for all $i$. Thus, $P^N$ is a row/right stochastic matrix.

Invariant measure estimation from transfer operator

The left invariant distribution $\mathbf{\rho}^N$ is a row vector, where $\mathbf{\rho}^N P^{N} = \mathbf{\rho}^N$. Hence, $\mathbf{\rho}^N$ is a row eigenvector of the transfer matrix $P^{N}$ associated with eigenvalue 1. The distribution $\mathbf{\rho}^N$ approximates the invariant density of the system subject to binning, and can be taken as a probability distribution over the partition elements.

In practice, the invariant measure $\mathbf{\rho}^N$ is computed using invariantmeasure, which also approximates the transfer matrix. The invariant distribution is initialized as a length-N random distribution which is then applied to $P^{N}$. The resulting length-N distribution is then applied to $P^{N}$ again. This process repeats until the difference between the distributions over consecutive iterations is below some threshold.

See also: RectangularBinning, invariantmeasure.

Utility methods/types

Entropies.InvariantMeasure — Type

InvariantMeasure(to, ρ)

Minimal return struct for invariantmeasure that contains the estimated invariant measure ρ, as well as the transfer operator to from which it is computed (including bin information).

See also: invariantmeasure.

Entropies.invariantmeasure — Function

invariantmeasure(x::AbstractDataset, binning::RectangularBinning) → iv::InvariantMeasure

Estimate an invariant measure over the points in x based on binning the data into rectangular boxes dictated by the binning, then approximate the transfer (Perron-Frobenius) operator over the bins. From the approximation to the transfer operator, compute an invariant distribution over the bins. Assumes that the input data are sequential.

Details on the estimation procedure is found the TransferOperator docstring.

Example

using DynamicalSystems, Plots, Entropies
D = 4
ds = Systems.lorenz96(D; F = 32.0)
N, dt = 20000, 0.1
orbit = trajectory(ds, N*dt; dt = dt, Ttr = 10.0)

# Estimate the invariant measure over some coarse graining of the orbit.
iv = invariantmeasure(orbit, RectangularBinning(15))

# Get the probabilities and bins
invariantmeasure(iv)

Probabilities and bin information

invariantmeasure(iv::InvariantMeasure) → (ρ::Probabilities, bins::Vector{<:SVector})

From a pre-computed invariant measure, return the probabilities and associated bins. The element ρ[i] is the probability of visitation to the box bins[i]. Analogous to binhist.

Transfer operator approach vs. naive histogram approach

Why bother with the transfer operator instead of using regular histograms to obtain probabilities?

In fact, the naive histogram approach and the transfer operator approach are equivalent in the limit of long enough time series (as $n \to \intfy$), which is guaranteed by the ergodic theorem. There is a crucial difference, however:

The naive histogram approach only gives the long-term probabilities that orbits visit a certain region of the state space. The transfer operator encodes that information too, but comes with the added benefit of knowing the transition probabilities between states (see transfermatrix).

See also: InvariantMeasure.

Entropies.transfermatrix — Function

transfermatrix(iv::InvariantMeasure) → (M::AbstractArray{<:Real, 2}, bins::Vector{<:SVector})

Return the transfer matrix/operator and corresponding bins. Here, bins[i] corresponds to the i-th row/column of the transfer matrix. Thus, the entry M[i, j] is the probability of jumping from the state defined by bins[i] to the state defined by bins[j].

See also: TransferOperator.

Kernel density

Entropies.NaiveKernel — Type

NaiveKernel(x, ϵ::Real; method = KDTree, w = 0, metric = Euclidean()) <: ProbabilitiesEstimator

Estimate probabilities/entropy using a "naive" kernel density estimation approach (KDE), as discussed in Prichard and Theiler (1995) [PrichardTheiler1995].

Probabilities $P(\mathbf{x}, \epsilon)$ are assigned to every point $\mathbf{x}$ by counting how many other points occupy the space spanned by a hypersphere of radius ϵ around $\mathbf{x}$, according to:

\[P_i( X, \epsilon) \approx \dfrac{1}{N} \sum_{s} B(||X_i - X_j|| < \epsilon),\]

where $B$ gives 1 if the argument is true. Probabilities are then normalized.

Keyword arguments

method = KDTree: the search structure supported by Neighborhood.jl. Specifically, use KDTree to use a tree-based neighbor search, or BruteForce for the direct distances between all points. KDTrees heavily outperform direct distances when the dimensionality of the data is much smaller than the data length.

w = 0: the Theiler window, which excludes indices $s$ that are within $|i - s| ≤ w$ from the given point $x_i$.

metric = Euclidean(): the distance metric.

Outcome space

The outcome space Ω for NaiveKernel are the indices of the input data, eachindex(x). The reason to not return the data points themselves is because duplicate data points may not have same probabilities (due to having different neighbors).

Timescales

Entropies.WaveletOverlap — Type

WaveletOverlap(x [,wavelet]) <: ProbabilitiesEstimator

Apply the maximal overlap discrete wavelet transform (MODWT) to a signal, then compute probabilities as the (normalized) energies at different wavelet scales. These probabilities are used to compute the wavelet entropy, according to Rosso et al. (2001)[Rosso2001]. Input timeseries x is needed for a well-defined outcome space.

By default the wavelet Wavelets.WT.Daubechies{12}() is used. Otherwise, you may choose a wavelet from the Wavelets package (it must subtype OrthoWaveletClass).

Outcome space

The outcome space for WaveletOverlap are the integers 1, 2, …, N enumerating the wavelet scales. To obtain a better understanding of what these mean, we prepared a notebook you can view online. As such, this estimator only works for timeseries input.

Entropies.PowerSpectrum — Type

PowerSpectrum(x_or_length(x)) <: ProbabilitiesEstimator

Calculate the power spectrum of a timeseries (amplitude square of its Fourier transform), and return the spectrum normalized to sum = 1 as probabilities. The Shannon entropy of these probabilities is typically referred in the literature as spectral entropy, e.g. [Llanos2016],[Tian2017].

The closer the spectrum is to flat, i.e., white noise, the higher the entropy. However, you can't compare entropies of timeseries with different length, because the binning in spectral space depends on the length of the input.

Outcome space

The outcome space Ω for PowerSpectrum is the set of frequencies in Fourier space. They should be multiplied with the sampling rate of the signal, which is assumed to be 1. The length of the input is therefore required for this estimator to have a well-defined outcome space.

Diversity

Entropies.Diversity — Type

Diversity(; m::Int, τ::Int, nbins::Int)

A ProbabilitiesEstimator based on the cosine similarity. It can be used with entropy to compute the diversity entropy of an input timeseries[Wang2020].

The implementation here allows for τ != 1, which was not considered in the original paper.

Description

Diversity probabilities are computed as follows.

From the input time series x, using embedding lag τ and embedding dimension m, construct the embedding $Y = \{\bf x_i \} = \{(x_{i}, x_{i+\tau}, x_{i+2\tau}, \ldots, x_{i+m\tau - 1}\}_{i = 1}^{N-mτ}$.

Compute $D = \{d(\bf x_t, \bf x_{t+1}) \}_{t=1}^{N-mτ-1}$, where $d(\cdot, \cdot)$ is the cosine similarity between two m-dimensional vectors in the embedding.

Divide the interval [-1, 1] into nbins equally sized subintervals.

Construct a histogram of cosine similarities $d \in D$ over those subintervals.

Sum-normalize the histogram to obtain probabilities.

Outcome space

The outcome space for Diversity is the bins of the [-1, 1] interval, and the return configuration is the same as in ValueHistogram (left bin edge).

BandtPompe2002Bandt, Christoph, and Bernd Pompe. "Permutation entropy: a natural complexity measure for time series." Physical review letters 88.17 (2002): 174102.

Zunino2017Zunino, L., Olivares, F., Scholkmann, F., & Rosso, O. A. (2017). Permutation entropy based time series analysis: Equalities in the input signal can lead to false conclusions. Physics Letters A, 381(22), 1883-1892.

He2016He, S., Sun, K., & Wang, H. (2016). Multivariate permutation entropy and its application for complexity analysis of chaotic systems. Physica A: Statistical Mechanics and its Applications, 461, 812-823.

Fadlallah2013Fadlallah, Bilal, et al. "Weighted-permutation entropy: A complexity measure for time series incorporating amplitude information." Physical Review E 87.2 (2013): 022911.

Azami2016Azami, H., & Escudero, J. (2016). Amplitude-aware permutation entropy: Illustration in spike detection and signal segmentation. Computer methods and programs in biomedicine, 128, 40-51.

Ribeiro2012Ribeiro et al. (2012). Complexity-entropy causality plane as a complexity measure for two-dimensional patterns. https://doi.org/10.1371/journal.pone.0040689

Schlemmer2018Schlemmer et al. (2018). Spatiotemporal Permutation Entropy as a Measure for Complexity of Cardiac Arrhythmia. https://doi.org/10.3389/fphy.2018.00039

Rostaghi2016Rostaghi, M., & Azami, H. (2016). Dispersion entropy: A measure for time-series analysis. IEEE Signal Processing Letters, 23(5), 610-614.

Li2018Li, G., Guan, Q., & Yang, H. (2018). Noise reduction method of underwater acoustic signals based on CEEMDAN, effort-to-compress complexity, refined composite multiscale dispersion entropy and wavelet threshold denoising. Entropy, 21(1), 11.

Azami2019Azami, H., da Silva, L. E. V., Omoto, A. C. M., & Humeau-Heurtier, A. (2019). Two-dimensional dispersion entropy: An information-theoretic method for irregularity analysis of images. Signal Processing: Image Communication, 75, 178-187.

Diego2019Diego, D., Haaga, K. A., & Hannisdal, B. (2019). Transfer entropy computation using the Perron-Frobenius operator. Physical Review E, 99(4), 042212.

PrichardTheiler1995Prichard, D., & Theiler, J. (1995). Generalized redundancies for time series analysis. Physica D: Nonlinear Phenomena, 84(3-4), 476-493.

Rosso2001Rosso et al. (2001). Wavelet entropy: a new tool for analysis of short duration brain electrical signals. Journal of neuroscience methods, 105(1), 65-75.

Llanos2016Llanos et al., Power spectral entropy as an information-theoretic correlate of manner of articulation in American English, The Journal of the Acoustical Society of America 141, EL127 (2017)

Tian2017Tian et al, Spectral Entropy Can Predict Changes of Working Memory Performance Reduced by Short-Time Training in the Delayed-Match-to-Sample Task, Front. Hum. Neurosci.

Wang2020Wang, X., Si, S., & Li, Y. (2020). Multiscale diversity entropy: A novel dynamical measure for fault diagnosis of rotating machinery. IEEE Transactions on Industrial Informatics, 17(8), 5419-5429.