Utility methods

Fast histograms

Missing docstring.

Missing docstring for Entropies.binhist. Check Documenter's build log for details.

Symbolization

Entropies.symbolize — Function

symbolize(x, scheme::SymbolizationScheme) → Vector{Int}
symbolize!(s, x, scheme::SymbolizationScheme) → Vector{Int}

Symbolize x using the provided symbolization scheme, optionally writing symbols into the pre-allocated symbol vector s if the scheme allows for it. For usage examples, see individual symbolization scheme docstrings.

The following symbolization schemes are currently implemented:

OrdinalPattern.

GaussianSymbolization.

RectangularBinEncoder.

Symbolization schemes

Entropies.GaussianSymbolization — Type

GaussianSymbolization(; c::Int = 3)

A symbolization scheme where the elements of x are symbolized into c distinct integer categories using the normal cumulative distribution function (NCDF).

Algorithm

Assume we have a univariate time series $X = \{x_i\}_{i=1}^N$. GaussianSymbolization first maps each $x_i$ to a new real number $y_i \in [0, 1]$ by using the normal cumulative distribution function (CDF), $x_i \to y_i : y_i = \dfrac{1}{ \sigma \sqrt{2 \pi}} \int_{-\infty}^{x_i} e^{(-(x_i - \mu)^2)/(2 \sigma^2)} dx$, where $\mu$ and $\sigma$ are the empirical mean and standard deviation of $X$.

Next, each $y_i$ is linearly mapped to an integer $z_i \in [1, 2, \ldots, c]$ using the map $y_i \to z_i : z_i = R(y_i(c-1) + 0.5)$, where $R$ indicates rounding up to the nearest integer. This procedure subdivides the interval $[0, 1]$ into $c$ different subintervals that form a covering of $[0, 1]$, and assigns each $y_i$ to one of these subintervals. The original time series $X$ is thus transformed to a symbol time series $S = \{ s_i \}_{i=1}^N$, where $s_i \in [1, 2, \ldots, c]$.

Usage

symbolize(x::AbstractVector, s::GaussianSymbolization)

Map the elements of x to a symbol time series according to the Gaussian symbolization scheme s.

Examples

julia> x = [0.1, 0.4, 0.7, -2.1, 8.0, 0.9, -5.2];

julia> Entropies.symbolize(x, GaussianSymbolization(c = 5))
7-element Vector{Int64}:
 3
 3
 3
 2
 5
 3
 1

See also: symbolize.

Entropies.OrdinalPattern — Type

OrdinalPattern(m = 3, τ = 1; lt = est.lt)

A symbolization scheme that converts the input time series to ordinal patterns, which are then encoded to integers using encode_motif.

Note

OrdinalPattern is intended for symbolizing time series. If providing a short vector, say x = [2, 5, 2, 1, 3, 4], then symbolize(x, OrdinalPattern(m = 2, τ = 1) will first embed x, then encode/symbolize each resulting state vector, not the original input. For symbolizing a single vector, use sortperm on it and use encode_motif on the resulting permutation indices.

Usage

symbolize(x, scheme::OrdinalPattern) → Vector{Int}
symbolize!(s, x, scheme::OrdinalPattern) → Vector{Int}

If applied to an m-dimensional Dataset x, then m and τ are ignored, and m-dimensional permutation patterns are obtained directly for each xᵢ ∈ x. Permutation patterns are then encoded as integers using encode_motif. Optionally, symbols can be written directly into a pre-allocated integer vector s, where length(s) == length(x) using symbolize!.

If applied to a univariate vector x, then x is first converted to a delay reconstruction using embedding dimension m and lag τ. Permutation patterns are then computed for each of the resulting m-dimensional xᵢ ∈ x, and each permutation is then encoded as an integer using encode_motif. If using the in-place variant with univariate input, s must obey length(s) == length(x)-(est.m-1)*est.τ.

Examples

using DelayEmbeddings, Entropies
D = Dataset([rand(7) for i = 1:1000])
s = symbolize(D, OrdinalPattern())

See also: symbolize.

Entropies.RectangularBinEncoder — Type

RectangularBinEncoder(x, binning::RectangularBinning) <: SymbolizationScheme

Find the minima along each dimension, and compute appropriate edge lengths for each dimension of x given a rectangular binning. Put them in an RectangularBinEncoder that can be then used to map points into bins via symbolize.

Encoding ordinal patterns

Entropies.encode_motif — Function

encode_motif(x, m::Int = length(x)) → s::Int

Encode the length-m motif x (a vector of indices that would sort some vector v in ascending order) into its unique integer symbol $s \in \{1, 2, \ldots, m - 1 \}$, using Algorithm 1 in Berger et al. (2019)[Berger2019].

Example

v = rand(5)

# The indices that would sort `v` in ascending order. This is now a permutation
# of the index permutation (1, 2, ..., 5)
x = sortperm(v)

# Encode this permutation as an integer.
encode_motif(x)

Alphabet length

Entropies.alphabet_length — Function

alphabet_length(x::Array_or_Dataset, est::ProbabilitiesEstimator) → Int

Return the total number of possible symbols/states implied by estimator for a given x. For some estimators, this total number is independent of x, in which case the input x is ignored and the method alphabet_length(est) is called.

If the total number of states cannot be known a priori, an error is thrown. Primarily used in entropy_normalized.

Examples

julia> est = SymbolicPermutation(m = 4);

julia> alphabet_length(rand(42), est) # same as `factorial(m)` for any `x`
24

Berger2019Berger, Sebastian, et al. "Teaching Ordinal Patterns to a Computer: Efficient Encoding Algorithms Based on the Lehmer Code." Entropy 21.10 (2019): 1023.